Retrieve a data frame of AUC scores

The auc function takes an S3 object generated by evalmod and retrieves a data frame with the Area Under the Curve (AUC) scores of ROC and Precision-Recall curves.

auc(curves)

# S3 method for aucs
auc(curves)

Arguments

curves

An S3 object generated by evalmod. The auc function accepts the following S3 objects.

S3 object 	 # of models

# of test datasets 	 sscurves

single 	 single

mscurves 	 multiple

single 	 smcurves

single 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with AUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. pauc for retieving a dataset of pAUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Shows AUCs
auc(sscurves)

#>   modnames dsids curvetypes      aucs
#> 1       m1     1        ROC 0.7200000
#> 2       m1     1        PRC 0.7397716

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Shows AUCs
auc(mscurves)

#>    modnames dsids curvetypes      aucs
#> 1    random     1        ROC 0.5507000
#> 2    random     1        PRC 0.5282579
#> 3   poor_er     1        ROC 0.8031000
#> 4   poor_er     1        PRC 0.7440634
#> 5   good_er     1        ROC 0.7532000
#> 6   good_er     1        PRC 0.7830621
#> 7     excel     1        ROC 0.9899000
#> 8     excel     1        PRC 0.9890602
#> 9      perf     1        ROC 1.0000000
#> 10     perf     1        PRC 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(10, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
sm_aucs <- auc(smcurves)

## Get AUCs of Precision-Recall
sm_aucs_prc <- subset(sm_aucs, curvetypes == "PRC")

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(10, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
mm_aucs <- auc(mmcurves)

## Get AUCs of Precision-Recall
mm_aucs_prc <- subset(mm_aucs, curvetypes == "PRC")