Retrieve a data frame of AUC scores

The auc function takes an S3 object generated by evalmod and retrieves a data frame with the Area Under the Curve (AUC) scores of ROC and Precision-Recall curves.

auc(curves)

# S3 method for aucs
auc(curves)

Arguments

curves

An S3 object generated by evalmod. The auc function accepts the following S3 objects.

S3 object 	 # of models

# of test datasets 	 sscurves

single 	 single

mscurves 	 multiple

single 	 smcurves

single 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with AUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. pauc for retieving a dataset of pAUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Shows AUCs
auc(sscurves)

#>   modnames dsids curvetypes      aucs
#> 1       m1     1        ROC 0.7200000
#> 2       m1     1        PRC 0.7397716

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Shows AUCs
auc(mscurves)

#>    modnames dsids curvetypes      aucs
#> 1    random     1        ROC 0.5507000
#> 2    random     1        PRC 0.5282579
#> 3   poor_er     1        ROC 0.8031000
#> 4   poor_er     1        PRC 0.7440634
#> 5   good_er     1        ROC 0.7532000
#> 6   good_er     1        PRC 0.7830621
#> 7     excel     1        ROC 0.9899000
#> 8     excel     1        PRC 0.9890602
#> 9      perf     1        ROC 1.0000000
#> 10     perf     1        PRC 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
sm_aucs <- auc(smcurves)

## Shows AUCs
sm_aucs

#>   modnames dsids curvetypes      aucs
#> 1  good_er     1        ROC 0.8174000
#> 2  good_er     1        PRC 0.8503615
#> 3  good_er     2        ROC 0.8293000
#> 4  good_er     2        PRC 0.8619951
#> 5  good_er     3        ROC 0.8239000
#> 6  good_er     3        PRC 0.8564469
#> 7  good_er     4        ROC 0.8245000
#> 8  good_er     4        PRC 0.8548463

## Get AUCs of Precision-Recall
sm_aucs_prc <- subset(sm_aucs, curvetypes == "PRC")

## Shows AUCs
sm_aucs_prc

#>   modnames dsids curvetypes      aucs
#> 2  good_er     1        PRC 0.8503615
#> 4  good_er     2        PRC 0.8619951
#> 6  good_er     3        PRC 0.8564469
#> 8  good_er     4        PRC 0.8548463

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
mm_aucs <- auc(mmcurves)

## Shows AUCs
mm_aucs

#>    modnames dsids curvetypes      aucs
#> 1    random     1        ROC 0.4656000
#> 2    random     1        PRC 0.4668442
#> 3   poor_er     1        ROC 0.7600000
#> 4   poor_er     1        PRC 0.7326417
#> 5   good_er     1        ROC 0.8100000
#> 6   good_er     1        PRC 0.8553404
#> 7     excel     1        ROC 0.9874000
#> 8     excel     1        PRC 0.9883639
#> 9      perf     1        ROC 1.0000000
#> 10     perf     1        PRC 1.0000000
#> 11   random     2        ROC 0.4662000
#> 12   random     2        PRC 0.4920597
#> 13  poor_er     2        ROC 0.8127000
#> 14  poor_er     2        PRC 0.7646622
#> 15  good_er     2        ROC 0.8037000
#> 16  good_er     2        PRC 0.8331458
#> 17    excel     2        ROC 0.9798000
#> 18    excel     2        PRC 0.9823248
#> 19     perf     2        ROC 1.0000000
#> 20     perf     2        PRC 1.0000000
#> 21   random     3        ROC 0.4459000
#> 22   random     3        PRC 0.4807934
#> 23  poor_er     3        ROC 0.8085000
#> 24  poor_er     3        PRC 0.7613053
#> 25  good_er     3        ROC 0.8203000
#> 26  good_er     3        PRC 0.8582972
#> 27    excel     3        ROC 0.9924000
#> 28    excel     3        PRC 0.9925871
#> 29     perf     3        ROC 1.0000000
#> 30     perf     3        PRC 1.0000000
#> 31   random     4        ROC 0.4985000
#> 32   random     4        PRC 0.4922140
#> 33  poor_er     4        ROC 0.7757000
#> 34  poor_er     4        PRC 0.7030577
#> 35  good_er     4        ROC 0.8211000
#> 36  good_er     4        PRC 0.8474317
#> 37    excel     4        ROC 0.9842000
#> 38    excel     4        PRC 0.9838969
#> 39     perf     4        ROC 1.0000000
#> 40     perf     4        PRC 1.0000000

## Get AUCs of Precision-Recall
mm_aucs_prc <- subset(mm_aucs, curvetypes == "PRC")

## Shows AUCs
mm_aucs_prc

#>    modnames dsids curvetypes      aucs
#> 2    random     1        PRC 0.4668442
#> 4   poor_er     1        PRC 0.7326417
#> 6   good_er     1        PRC 0.8553404
#> 8     excel     1        PRC 0.9883639
#> 10     perf     1        PRC 1.0000000
#> 12   random     2        PRC 0.4920597
#> 14  poor_er     2        PRC 0.7646622
#> 16  good_er     2        PRC 0.8331458
#> 18    excel     2        PRC 0.9823248
#> 20     perf     2        PRC 1.0000000
#> 22   random     3        PRC 0.4807934
#> 24  poor_er     3        PRC 0.7613053
#> 26  good_er     3        PRC 0.8582972
#> 28    excel     3        PRC 0.9925871
#> 30     perf     3        PRC 1.0000000
#> 32   random     4        PRC 0.4922140
#> 34  poor_er     4        PRC 0.7030577
#> 36  good_er     4        PRC 0.8474317
#> 38    excel     4        PRC 0.9838969
#> 40     perf     4        PRC 1.0000000