Retrieve a data frame of AUC scores

The auc function takes an S3 object generated by evalmod and retrieves a data frame with the Area Under the Curve (AUC) scores of ROC and Precision-Recall curves.

auc(curves)

# S3 method for aucs
auc(curves)

Arguments

curves

An S3 object generated by evalmod. The auc function accepts the following S3 objects.

S3 object	# of models	# of test datasets

sscurves	single	single

mscurves	multiple	single

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with AUC scores.

See also

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Shows AUCs
auc(sscurves)
#>   modnames dsids curvetypes      aucs
#> 1       m1     1        ROC 0.7200000
#> 2       m1     1        PRC 0.7397716

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]]
)

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Shows AUCs
auc(mscurves)
#>    modnames dsids curvetypes      aucs
#> 1    random     1        ROC 0.4546000
#> 2    random     1        PRC 0.4671165
#> 3   poor_er     1        ROC 0.8238000
#> 4   poor_er     1        PRC 0.7966517
#> 5   good_er     1        ROC 0.7342000
#> 6   good_er     1        PRC 0.7763444
#> 7     excel     1        ROC 0.9884000
#> 8     excel     1        PRC 0.9885669
#> 9      perf     1        ROC 1.0000000
#> 10     perf     1        PRC 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
sm_aucs <- auc(smcurves)

## Shows AUCs
sm_aucs
#>   modnames dsids curvetypes      aucs
#> 1  good_er     1        ROC 0.8413000
#> 2  good_er     1        PRC 0.8702842
#> 3  good_er     2        ROC 0.8549000
#> 4  good_er     2        PRC 0.8860195
#> 5  good_er     3        ROC 0.8152000
#> 6  good_er     3        PRC 0.8454227
#> 7  good_er     4        ROC 0.8308000
#> 8  good_er     4        PRC 0.8590165

## Get AUCs of Precision-Recall
sm_aucs_prc <- subset(sm_aucs, curvetypes == "PRC")

## Shows AUCs
sm_aucs_prc
#>   modnames dsids curvetypes      aucs
#> 2  good_er     1        PRC 0.8702842
#> 4  good_er     2        PRC 0.8860195
#> 6  good_er     3        PRC 0.8454227
#> 8  good_er     4        PRC 0.8590165

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Get AUCs
mm_aucs <- auc(mmcurves)

## Shows AUCs
mm_aucs
#>    modnames dsids curvetypes      aucs
#> 1    random     1        ROC 0.5560000
#> 2    random     1        PRC 0.5484669
#> 3   poor_er     1        ROC 0.8120000
#> 4   poor_er     1        PRC 0.7383763
#> 5   good_er     1        ROC 0.7640000
#> 6   good_er     1        PRC 0.8122955
#> 7     excel     1        ROC 0.9796000
#> 8     excel     1        PRC 0.9801895
#> 9      perf     1        ROC 1.0000000
#> 10     perf     1        PRC 1.0000000
#> 11   random     2        ROC 0.4223000
#> 12   random     2        PRC 0.4339653
#> 13  poor_er     2        ROC 0.7994000
#> 14  poor_er     2        PRC 0.7207124
#> 15  good_er     2        ROC 0.7254000
#> 16  good_er     2        PRC 0.7858929
#> 17    excel     2        ROC 0.9829000
#> 18    excel     2        PRC 0.9856180
#> 19     perf     2        ROC 1.0000000
#> 20     perf     2        PRC 1.0000000
#> 21   random     3        ROC 0.4575000
#> 22   random     3        PRC 0.4746854
#> 23  poor_er     3        ROC 0.8060000
#> 24  poor_er     3        PRC 0.7424152
#> 25  good_er     3        ROC 0.7331000
#> 26  good_er     3        PRC 0.7865790
#> 27    excel     3        ROC 0.9720000
#> 28    excel     3        PRC 0.9734300
#> 29     perf     3        ROC 1.0000000
#> 30     perf     3        PRC 1.0000000
#> 31   random     4        ROC 0.4196000
#> 32   random     4        PRC 0.4630314
#> 33  poor_er     4        ROC 0.7424000
#> 34  poor_er     4        PRC 0.6645997
#> 35  good_er     4        ROC 0.7796000
#> 36  good_er     4        PRC 0.8176798
#> 37    excel     4        ROC 0.9786000
#> 38    excel     4        PRC 0.9770366
#> 39     perf     4        ROC 1.0000000
#> 40     perf     4        PRC 1.0000000

## Get AUCs of Precision-Recall
mm_aucs_prc <- subset(mm_aucs, curvetypes == "PRC")

## Shows AUCs
mm_aucs_prc
#>    modnames dsids curvetypes      aucs
#> 2    random     1        PRC 0.5484669
#> 4   poor_er     1        PRC 0.7383763
#> 6   good_er     1        PRC 0.8122955
#> 8     excel     1        PRC 0.9801895
#> 10     perf     1        PRC 1.0000000
#> 12   random     2        PRC 0.4339653
#> 14  poor_er     2        PRC 0.7207124
#> 16  good_er     2        PRC 0.7858929
#> 18    excel     2        PRC 0.9856180
#> 20     perf     2        PRC 1.0000000
#> 22   random     3        PRC 0.4746854
#> 24  poor_er     3        PRC 0.7424152
#> 26  good_er     3        PRC 0.7865790
#> 28    excel     3        PRC 0.9734300
#> 30     perf     3        PRC 1.0000000
#> 32   random     4        PRC 0.4630314
#> 34  poor_er     4        PRC 0.6645997
#> 36  good_er     4        PRC 0.8176798
#> 38    excel     4        PRC 0.9770366
#> 40     perf     4        PRC 1.0000000