Calculate CIs of ROC and precision-recall AUCs

The auc_ci function takes an S3 object generated by evalmod and calculates CIs of AUCs when multiple data sets are specified.

auc_ci(curves, alpha = NULL, dtype = NULL)

# S3 method for aucs
auc_ci(curves, alpha = 0.05, dtype = "normal")

Arguments

curves

An S3 object generated by evalmod. The auc_ci function accepts the following S3 objects.

S3 object	# of models	# of test datasets

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

alpha

A numeric value of the significant level (default: 0.05)

dtype

A string to specify the distribution used for CI calculation.

dtype	distribution

normal (default)	Normal distribution

z	Normal distribution

t	t-distribution

Value

The auc_ci function returns a dataframe of AUC CIs.

See also

Examples

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat)

## Calculate CI of AUCs
sm_auc_cis <- auc_ci(smcurves)

## Shows the result
sm_auc_cis
#>   modnames curvetypes     mean      error lower_bound upper_bound n
#> 1  good_er        ROC 0.801325 0.04192196   0.7594030   0.8432470 4
#> 2  good_er        PRC 0.840212 0.03524380   0.8049682   0.8754558 4

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat)

## Calculate CI of AUCs
mm_auc_ci <- auc_ci(mmcurves)

## Shows the result
mm_auc_ci
#>    modnames curvetypes      mean       error lower_bound upper_bound n
#> 1    random        ROC 0.4948500 0.027610054   0.4672399   0.5224601 4
#> 2    random        PRC 0.5119320 0.028831600   0.4831004   0.5407636 4
#> 3   poor_er        ROC 0.7825500 0.044777457   0.7377725   0.8273275 4
#> 4   poor_er        PRC 0.7225722 0.064276152   0.6582961   0.7868484 4
#> 5   good_er        ROC 0.7983250 0.011349014   0.7869760   0.8096740 4
#> 6   good_er        PRC 0.8328021 0.016809924   0.8159922   0.8496120 4
#> 7     excel        ROC 0.9788000 0.007728407   0.9710716   0.9865284 4
#> 8     excel        PRC 0.9803272 0.007219165   0.9731081   0.9875464 4
#> 9      perf        ROC 1.0000000 0.000000000   1.0000000   1.0000000 4
#> 10     perf        PRC 1.0000000 0.000000000   1.0000000   1.0000000 4