Calculate CIs of ROC and precision-recall AUCs

The auc_ci function takes an S3 object generated by evalmod and calculates CIs of AUCs when multiple data sets are specified.

auc_ci(curves, alpha = NULL, dtype = NULL)

# S3 method for aucs
auc_ci(curves, alpha = 0.05, dtype = "normal")

Arguments

curves

An S3 object generated by evalmod. The auc_ci function accepts the following S3 objects.

S3 object	# of models	# of test datasets

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

alpha

A numeric value of the significant level (default: 0.05)

dtype

A string to specify the distribution used for CI calculation.

dtype	distribution

normal (default)	Normal distribution

z	Normal distribution

t	t-distribution

Value

The auc_ci function returns a dataframe of AUC CIs.

See also

Examples

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat)

## Calculate CI of AUCs
sm_auc_cis <- auc_ci(smcurves)

## Shows the result
sm_auc_cis
#>   modnames curvetypes      mean      error lower_bound upper_bound n
#> 1  good_er        ROC 0.8055500 0.03170960   0.7738404   0.8372596 4
#> 2  good_er        PRC 0.8390638 0.01862613   0.8204377   0.8576900 4

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat)

## Calculate CI of AUCs
mm_auc_ci <- auc_ci(mmcurves)

## Shows the result
mm_auc_ci
#>    modnames curvetypes      mean       error lower_bound upper_bound n
#> 1    random        ROC 0.5061500 0.028297846   0.4778522   0.5344478 4
#> 2    random        PRC 0.5056945 0.015950144   0.4897444   0.5216447 4
#> 3   poor_er        ROC 0.7904500 0.031876340   0.7585737   0.8223263 4
#> 4   poor_er        PRC 0.7311314 0.025387763   0.7057436   0.7565191 4
#> 5   good_er        ROC 0.7974500 0.029119870   0.7683301   0.8265699 4
#> 6   good_er        PRC 0.8322786 0.023196748   0.8090818   0.8554753 4
#> 7     excel        ROC 0.9860000 0.004673874   0.9813261   0.9906739 4
#> 8     excel        PRC 0.9851993 0.004973312   0.9802260   0.9901727 4
#> 9      perf        ROC 1.0000000 0.000000000   1.0000000   1.0000000 4
#> 10     perf        PRC 1.0000000 0.000000000   1.0000000   1.0000000 4