Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] 0.996 -0.807 -0.416 -1.688 -0.464 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.731 -0.65 1.408 0.179 0.565 ...
#>   .. ..$ : num [1:30] 0.658 0.581 0.985 0.841 0.938 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.0819 -0.2068 -0.1062 1.9851 0.4403 ...
#>   .. ..$ : num [1:30] 0.879 0.696 0.829 0.985 0.986 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.0575 0.0106 -0.0546 -0.5099 -0.7034 ...
#>   .. ..$ : num [1:10] 0.852 0.991 0.928 0.933 0.5 ...
#>   .. ..$ : num [1:10] 0.175 0.483 0.505 0.714 0.412 ...
#>   .. ..$ : num [1:10] 2.77 3.95 3.61 2.35 3.23 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.934 -0.164 1.479 -0.394 -0.478 ...
#>   .. ..$ : num [1:10] 0.718 0.527 0.814 0.902 0.996 ...
#>   .. ..$ : num [1:10] 0.915 0.62 0.827 0.553 0.874 ...
#>   .. ..$ : num [1:10] 3.2 1.71 4.02 1.04 3.67 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.391 0.211 0.325 0.303 0.709 ...
#>   .. ..$ : num [1:10] 0.914 0.996 0.954 0.971 0.481 ...
#>   .. ..$ : num [1:10] 0.551 0.457 0.335 0.445 0.912 ...
#>   .. ..$ : num [1:10] 3.47 3.15 3.25 3.28 4.47 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...