Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] 1.636 -0.873 -0.549 1.123 1.537 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 1.857 1.932 -0.478 0.827 1.46 ...
#>   .. ..$ : num [1:30] 0.94 0.977 0.978 0.907 0.741 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -1.104 0.977 0.909 1.608 -1.008 ...
#>   .. ..$ : num [1:30] 0.989 0.302 0.45 0.495 0.79 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.715 1.897 -1.164 -0.507 0.382 ...
#>   .. ..$ : num [1:10] 0.956 0.936 1 0.89 0.839 ...
#>   .. ..$ : num [1:10] 0.537 0.314 0.23 0.618 0.291 ...
#>   .. ..$ : num [1:10] 2.05 2.031 1.241 2.05 0.898 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.00475 1.16552 0.09601 1.97457 0.80156 ...
#>   .. ..$ : num [1:10] 0.665 0.657 0.359 0.992 0.802 ...
#>   .. ..$ : num [1:10] 0.0628 0.8346 0.3095 0.9849 0.3868 ...
#>   .. ..$ : num [1:10] 3.4 3.75 2.51 4.49 3.54 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.0995 0.2421 0.2417 0.4908 0.9295 ...
#>   .. ..$ : num [1:10] 0.974 0.892 0.985 0.858 0.989 ...
#>   .. ..$ : num [1:10] 0.7335 0.9708 0.2612 0.0134 0.5109 ...
#>   .. ..$ : num [1:10] 3.75 4.33 3.43 2.7 2.76 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...