Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] 0.227 -1.094 -0.336 0.144 0.584 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.795 0.465 -0.37 0.209 0.288 ...
#>   .. ..$ : num [1:30] 0.839 0.999 0.792 0.636 0.989 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 0.36 0.965 -1.817 -0.298 -0.127 ...
#>   .. ..$ : num [1:30] 0.953 0.82 0.945 0.952 0.788 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.353 1.102 -2.034 1.629 0.659 ...
#>   .. ..$ : num [1:10] 0.968 0.953 0.674 0.873 0.716 ...
#>   .. ..$ : num [1:10] 0.835 0.79 0.145 0.713 0.755 ...
#>   .. ..$ : num [1:10] 2.87 3.2 3.2 2.98 3.19 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.45 -1.638 1.256 0.499 0.419 ...
#>   .. ..$ : num [1:10] 0.827 0.851 0.985 0.835 0.442 ...
#>   .. ..$ : num [1:10] 0.492 0.128 0.941 0.522 0.608 ...
#>   .. ..$ : num [1:10] 2.84 3.37 3.03 3.22 2.81 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -1.49492 2.50508 0.56481 0.75014 -0.00288 ...
#>   .. ..$ : num [1:10] 0.614 0.702 0.937 0.952 0.926 ...
#>   .. ..$ : num [1:10] 0.408 0.757 0.712 0.827 0.439 ...
#>   .. ..$ : num [1:10] 3.15 2.24 2.9 2.31 3.08 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...