Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] -1.693 -0.384 0.981 -0.437 -0.75 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -1.011 0.369 2.673 -1.004 -0.94 ...
#>   .. ..$ : num [1:30] 0.949 0.835 0.968 0.764 0.657 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.751 0.499 0.982 1.109 -0.332 ...
#>   .. ..$ : num [1:30] 0.925 0.741 0.994 0.971 0.979 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 1.145 -0.761 0.363 -0.849 1.31 ...
#>   .. ..$ : num [1:10] 0.928 0.577 0.93 0.862 0.975 ...
#>   .. ..$ : num [1:10] 0.805 0.385 0.339 0.685 0.619 ...
#>   .. ..$ : num [1:10] 2.64 3.03 3.91 4.24 2.98 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.0291 1.1257 0.4641 -0.5914 1.0958 ...
#>   .. ..$ : num [1:10] 0.522 0.743 0.818 0.929 0.815 ...
#>   .. ..$ : num [1:10] 0.177 0.846 0.453 0.034 0.635 ...
#>   .. ..$ : num [1:10] 3.38 3.41 3.37 2.71 2.63 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.91 1.817 0.197 -0.668 -1.349 ...
#>   .. ..$ : num [1:10] 0.648 0.548 0.821 0.774 0.863 ...
#>   .. ..$ : num [1:10] 0.799 0.958 0.522 0.569 0.085 ...
#>   .. ..$ : num [1:10] 2.35 2.85 4.21 5.73 3.15 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...