Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat
The number of iterations to make samples.

np
The number of positives in a sample.

nn
The number of negatives in a sample.

score_names
A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)

#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] 1.865 -1.412 0.171 -0.345 0.192 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)

#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.6 1.468 -0.62 0.567 1.385 ...
#>   .. ..$ : num [1:30] 0.355 0.438 0.724 0.946 0.842 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 0.1267 0.0642 0.6424 0.0919 1.0682 ...
#>   .. ..$ : num [1:30] 0.55 0.823 0.853 0.947 0.823 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)

#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -1.1942 -0.8237 0.9838 1.566 0.0735 ...
#>   .. ..$ : num [1:10] 0.906 0.76 0.902 0.896 0.666 ...
#>   .. ..$ : num [1:10] 0.607 0.402 0.43 0.935 0.816 ...
#>   .. ..$ : num [1:10] 3.14 2.41 4.7 3.42 3.59 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.379 -0.195 -1.483 -0.873 2.221 ...
#>   .. ..$ : num [1:10] 0.947 0.737 0.893 0.81 0.88 ...
#>   .. ..$ : num [1:10] 0.4308 0.0707 0.7081 0.8363 0.5821 ...
#>   .. ..$ : num [1:10] 3.51 3.12 3.35 3.51 5.69 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 1.659 0.332 -0.548 -1.313 1.668 ...
#>   .. ..$ : num [1:10] 0.941 0.887 0.643 0.992 0.96 ...
#>   .. ..$ : num [1:10] 0.881 0.584 0.561 0.658 0.542 ...
#>   .. ..$ : num [1:10] 2.44 4.34 4.15 2.63 2.96 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...