Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] 0.8418 -0.7251 1.5303 1.6161 0.0609 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.634 -0.656 -0.9 -0.187 0.625 ...
#>   .. ..$ : num [1:30] 0.863 0.806 0.708 0.911 0.668 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.5875 0.9386 -0.4697 0.0601 -2.7927 ...
#>   .. ..$ : num [1:30] 0.791 0.767 0.888 0.983 0.943 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.929 -1.384 0.232 -0.265 1.568 ...
#>   .. ..$ : num [1:10] 0.907 0.845 0.867 0.91 0.869 ...
#>   .. ..$ : num [1:10] 0.263 0.978 0.646 0.769 0.455 ...
#>   .. ..$ : num [1:10] 3.23 2.26 1.79 1.77 2.71 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.762 -0.328 0.437 -0.997 0.767 ...
#>   .. ..$ : num [1:10] 0.601 0.96 0.805 0.961 0.834 ...
#>   .. ..$ : num [1:10] 0.515 0.094 0.946 0.859 0.481 ...
#>   .. ..$ : num [1:10] 2.35 3.02 2.55 2.89 5.2 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -1.24384 -0.97327 0.20168 0.00278 -0.52765 ...
#>   .. ..$ : num [1:10] 0.709 0.914 0.614 0.843 0.953 ...
#>   .. ..$ : num [1:10] 0.8247 0.0276 0.0992 0.5589 0.7199 ...
#>   .. ..$ : num [1:10] 3.13 1.26 2.78 2.21 3.17 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...