Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat
The number of iterations to make samples.

np
The number of positives in a sample.

nn
The number of negatives in a sample.

score_names
A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

mmdata for formatting input data. evalmod for calculation evaluation measures.

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)

#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] -0.0129 -0.2613 -0.2291 -0.2158 -1.5068 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)

#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -1.423 -0.859 1.727 0.343 -0.818 ...
#>   .. ..$ : num [1:30] 0.943 0.863 0.837 0.899 0.857 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 0.174 -2.387 -0.949 -0.653 0.308 ...
#>   .. ..$ : num [1:30] 0.835 0.993 0.737 0.555 0.703 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)

#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.155 1.03 -0.862 -1.68 -1.126 ...
#>   .. ..$ : num [1:10] 0.535 0.546 0.692 0.85 0.942 ...
#>   .. ..$ : num [1:10] 0.465 0.946 0.852 0.253 0.452 ...
#>   .. ..$ : num [1:10] 1.35 2.14 2.47 3.03 1.32 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 1.269 -1.441 0.401 0.945 -0.554 ...
#>   .. ..$ : num [1:10] 0.97 0.645 0.875 0.97 0.906 ...
#>   .. ..$ : num [1:10] 0.61739 0.00383 0.31684 0.88459 0.4428 ...
#>   .. ..$ : num [1:10] 0.944 4.944 0.475 3.166 2.56 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.0391 1.5977 1.1102 -0.6597 1.6793 ...
#>   .. ..$ : num [1:10] 0.986 0.838 0.976 0.676 0.907 ...
#>   .. ..$ : num [1:10] 0.129 0.403 0.105 0.605 0.79 ...
#>   .. ..$ : num [1:10] 5.06 2.69 1.32 1.14 1.89 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...