Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] -0.265 0.249 -1.593 -0.771 -0.167 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.933 0.799 -0.538 0.715 2.01 ...
#>   .. ..$ : num [1:30] 0.998 0.827 0.607 0.698 0.828 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 0.529 0.523 0.541 1.237 -1.027 ...
#>   .. ..$ : num [1:30] 0.712 0.876 0.767 0.93 0.722 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.8782 -0.0969 -0.1893 1.6993 0.5599 ...
#>   .. ..$ : num [1:10] 0.838 0.997 0.781 0.81 0.928 ...
#>   .. ..$ : num [1:10] 0.513 0.506 0.299 0.112 0.575 ...
#>   .. ..$ : num [1:10] 2.05 3.63 2.71 4.47 2.27 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -1.31 -0.455 -0.672 -1.717 0.609 ...
#>   .. ..$ : num [1:10] 0.981 0.785 0.769 0.395 0.402 ...
#>   .. ..$ : num [1:10] 0.661 0.597 0.407 0.1 0.347 ...
#>   .. ..$ : num [1:10] 2.53 1.3 3.35 2.06 2.53 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 2.5861 -0.0256 1.1895 0.4456 0.2854 ...
#>   .. ..$ : num [1:10] 0.683 0.791 0.685 0.28 0.948 ...
#>   .. ..$ : num [1:10] 0.11 0.942 0.915 0.94 0.671 ...
#>   .. ..$ : num [1:10] 1.67 2.53 3.77 3.97 2.76 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...