Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)
#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] -0.722 0.244 -0.286 0.659 1.687 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)
#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -0.7088 -0.0982 -1.4427 -0.1644 1.03 ...
#>   .. ..$ : num [1:30] 0.632 0.829 0.726 0.879 0.676 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 0.742 0.184 -0.619 0.76 -1.72 ...
#>   .. ..$ : num [1:30] 0.837 0.572 0.918 0.763 0.942 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)
#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.0409 0.1895 1.0399 0.3136 0.2282 ...
#>   .. ..$ : num [1:10] 0.824 0.962 0.988 0.86 0.87 ...
#>   .. ..$ : num [1:10] 0.884 0.565 0.71 0.365 0.486 ...
#>   .. ..$ : num [1:10] 2.74 3.38 3.57 3.49 3.87 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.232 -0.411 -2.232 -0.713 1.703 ...
#>   .. ..$ : num [1:10] 0.878 0.978 0.97 0.99 0.479 ...
#>   .. ..$ : num [1:10] 0.989 0.457 0.881 0.597 0.547 ...
#>   .. ..$ : num [1:10] 2.86 3.53 3.57 4.54 2.75 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] 0.0371 1.5116 -0.2915 -0.847 -1.7812 ...
#>   .. ..$ : num [1:10] 0.914 0.996 0.846 0.828 0.983 ...
#>   .. ..$ : num [1:10] 0.8191 0.0779 0.6361 0.6051 0.877 ...
#>   .. ..$ : num [1:10] 4.24 1.26 2.68 4.92 3.45 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...