Create random samples for simulations

The create_sim_samples function generates random samples with different performance levels.

create_sim_samples(n_repeat, np, nn, score_names = "random")

Arguments

n_repeat

The number of iterations to make samples.

np

The number of positives in a sample.

nn

The number of negatives in a sample.

score_names

A character vector for the names of the following performance levels.

"random"

Random

"poor_er"

Poor early retrieval

"good_er"

Good early retrieval

"excel"

Excellent

"perf"

Perfect

"all"

All of the above

Value

The create_sim_samples function returns a list with the following items.

scores: a list of numeric vectors

labels: an integer vector

modnames: a character vector of the model names

dsids: a character vector of the dataset IDs

See also

mmdata for formatting input data. evalmod for calculation evaluation measures.

Examples

##################################################
### Create a set of samples with 10 positives and 10 negatives
### for the random performance level
###
samps1 <- create_sim_samples(1, 10, 10, "random")

## Show the list structure
str(samps1)

#> List of 4
#>  $ scores  :List of 1
#>   ..$ :List of 1
#>   .. ..$ : num [1:20] -0.5361 1.0082 -1.3957 0.0937 1.8347 ...
#>  $ labels  : num [1:20] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr "random"
#>  $ dsids   : int 1

##################################################
### Create two sets of samples with 10 positives and 20 negatives
### for the random and the poor early retrieval performance levels
###
samps2 <- create_sim_samples(2, 10, 20, c("random", "poor_er"))

## Show the list structure
str(samps2)

#> List of 4
#>  $ scores  :List of 2
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] -1.254 -0.667 -1.06 -0.736 -0.531 ...
#>   .. ..$ : num [1:30] 0.928 0.58 0.619 0.983 0.986 ...
#>   ..$ :List of 2
#>   .. ..$ : num [1:30] 1.267 0.784 0.136 0.717 1.261 ...
#>   .. ..$ : num [1:30] 0.995 0.936 0.679 0.844 0.647 ...
#>  $ labels  : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ modnames: chr [1:4] "random" "poor_er" "random" "poor_er"
#>  $ dsids   : int [1:4] 1 1 2 2

##################################################
### Create 3 sets of samples with 5 positives and 5 negatives
### for all 5 levels
###
samps3 <- create_sim_samples(3, 5, 5, "all")

## Show the list structure
str(samps3)

#> List of 4
#>  $ scores  :List of 3
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.393 1.128 -0.427 0.945 -1.528 ...
#>   .. ..$ : num [1:10] 0.715 0.739 0.869 0.945 0.695 ...
#>   .. ..$ : num [1:10] 0.342 0.829 0.534 0.791 0.412 ...
#>   .. ..$ : num [1:10] 2.54 3.77 4.71 4.13 3.89 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.531 0.7293 -0.5521 0.0318 -1.8913 ...
#>   .. ..$ : num [1:10] 0.819 0.901 0.74 0.542 0.995 ...
#>   .. ..$ : num [1:10] 0.944 0.401 0.743 0.786 0.392 ...
#>   .. ..$ : num [1:10] 2.47 2.39 2.06 3.11 1.58 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>   ..$ :List of 5
#>   .. ..$ : num [1:10] -0.822 -0.448 -1.451 -1.073 1.501 ...
#>   .. ..$ : num [1:10] 0.972 0.837 0.795 0.828 0.796 ...
#>   .. ..$ : num [1:10] 0.1501 0.7404 0.857 0.6114 0.0613 ...
#>   .. ..$ : num [1:10] 3.69 2.77 2.2 4.18 4.04 ...
#>   .. ..$ : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ labels  : num [1:10] 1 1 1 1 1 0 0 0 0 0
#>  $ modnames: chr [1:15] "random" "poor_er" "good_er" "excel" ...
#>  $ dsids   : int [1:15] 1 1 1 1 1 2 2 2 2 2 ...