Evaluate models and calculate performance evaluation measures

The evalmod function calculates ROC and Precision-Recall curves for specified prediction scores and binary labels. It also calculate several basic performance evaluation measures, such as accuracy, error rate, and precision, by specifying mode as "basic".

evalmod(mdat, mode = NULL, scores = NULL, labels = NULL,
  modnames = NULL, dsids = NULL, posclass = NULL, na_worst = TRUE,
  ties_method = "equiv", calc_avg = TRUE, cb_alpha = 0.05,
  raw_curves = FALSE, x_bins = 1000, ...)

Arguments

mdat
An S3 object created by the mmdata function. It contains formatted scores and labels. The evalmod function ignores the following arguments when mdat is specified.

scores

labels

modnames

dsids

posclass

na_worst

ties_method

These arguments are internally passed to the mmdata function when mdat is unspecified. In that case, both scores and labels must be at least specified.

mode
A string that specifies the types of evaluation measures that the evalmod function calculates.

"rocprc"

ROC and Precision-Recall curves

"prcroc"

Same as above

"basic"

Normalized ranks vs. accuracy, error rate, specificity, sensitivity, precision, Matthews correlation coefficient, and F-score.

"aucroc"

Fast AUC(ROC) calculation with the U statistic

scores
A numeric dataset of predicted scores. It can be a vector, a matrix, an array, a data frame, or a list. The join_scores function can be useful to make scores with multiple datasets.

labels
A numeric, character, logical, or factor dataset of observed labels. It can be a vector, a matrix, an array, a data frame, or a list. The join_labels function can be useful to make labels with multiple datasets.

modnames
A character vector for the names of the models. The evalmod function automatically generates default names as "m1", "m2", "m3", and so on when it is NULL.

dsids
A numeric vector for test dataset IDs. The evalmod function automatically generates the default ID as 1 when it is NULL.

posclass
A scalar value to specify the label of positives in labels. It must be the same data type as labels. For example, posclass = -1 changes the positive label from 1 to -1 when labels contains 1 and -1. The positive label will be automatically detected when posclass is NULL.

na_worst
A Boolean value for controlling the treatment of NAs in scores.

TRUE

All NAs are treated as the worst scores

FALSE

All NAs are treated as the best scores

ties_method
A string for controlling ties in scores.

"equiv"

Ties are equivalently ranked

"first"

Ties are ranked in an increasing order as appeared

"random"

Ties are ranked in random order

calc_avg
A logical value to specify whether average curves should be calculated. It is effective only when dsids contains multiple dataset IDs. For instance, the function calculates the average for the model "m1" when modnames is c("m1", "m1", "m1") and dsids is c(1, 2, 3). The calculation points are defined by x_bins.

cb_alpha
A numeric value with range [0, 1] to specify the alpha value of the point-wise confidence bounds calculation. It is effective only when calc_avg is set to TRUE. For example, it should be 0.05 for the 95% confidence level. The calculation points are defined by x_bins.

raw_curves
A logical value to specify whether all raw curves should be discarded after the average curves are calculated. It is effective only when calc_avg is set to TRUE.

x_bins
An integer value to specify the number of minimum bins on the x-axis. It is then used to define supporting poins. For instance, the x-values of the supporting points will be c(0, 0.5, 1) and c(0, 0.25, 0.5, 0.75, 1) when x_bins = 2 and x_bins = 4, respectively. All corresponding y-values of the supporting points are calculated.

...
These additional arguments are passed to mmdata for data preparation.

Value

The evalmod function returns an S3 object that contains performance evaluation measures. The number of models and the number of datasets can be controlled by modnames and dsids. For example, the number of models is "single" and the number of test datasets is "multiple" when modnames = c("m1", "m1", "m1") and dsids = c(1, 2, 3) are specified.

Different S3 objects have different default behaviors of S3 generics, such as plot, autoplot, and fortify.

The evalmod function returns one of the following S3 objects when mode is "prcroc". The objects contain ROC and Precision-Recall curves.

S3 object 	# of models 	# of test datasets

sscurves 	 single 	 single

mscurves 	 multiple 	 single

smcurves 	 single 	 multiple

mmcurves 	 multiple 	 multiple

The evalmod function returns one of the following S3 objects when mode is "basic". They contain five different basic evaluation measures; error rate, accuracy, specificity, sensitivity, and precision.

S3 object 	# of models 	# of test datasets

sspoints 	 single 	 single

mspoints 	 multiple 	 single

smpoints 	 single 	 multiple

mmpoints 	 multiple 	 multiple

The evalmod function returns the aucroc S3 object when mode is "aucroc", which can be used with 'print' and 'as.data.frame'.

See also

plot for plotting curves with the general R plot. autoplot and fortify for plotting curves with ggplot2. mmdata for formatting input data. join_scores and join_labels for formatting scores and labels with multiple datasets. format_nfold for creating n-fold cross validation dataset from data frame. create_sim_samples for generating random samples for simulations.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)
sscurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1         m1          1        ROC 0.7200000
#>    2         m1          1        PRC 0.7397716
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1             10             10
#> 

## Generate an sspoints object that contains basic evaluation measures
sspoints <- evalmod(mode = "basic", scores = P10N10$scores,
                    labels = P10N10$labels)
sspoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>      Model ID  Meas.       Min.    1st Qu.     Median       Mean    3rd Qu.
#>    1    m1  1   rank  0.0000000  0.2500000  0.5000000  0.5000000  0.7500000
#>    2    m1  1  score  5.0000000  5.7500000 14.0000000 11.7500000 15.2500000
#>    3    m1  1  label -1.0000000 -1.0000000  0.0000000  0.0000000  1.0000000
#>    4    m1  1    err  0.3000000  0.3500000  0.4000000  0.3952381  0.4400000
#>    5    m1  1    acc  0.5000000  0.5600000  0.6000000  0.6047619  0.6500000
#>    6    m1  1     sp  0.0000000  0.4000000  0.6333333  0.6047619  0.9000000
#>    7    m1  1     sn  0.0000000  0.4000000  0.6333333  0.6047619  0.9000000
#>    8    m1  1   prec  0.5000000  0.5750000  0.6333333  0.6892147  0.7619048
#>    9    m1  1    mcc  0.1376494  0.2238168  0.2666667  0.2755698  0.3367701
#>   10    m1  1 fscore  0.0000000  0.5333333  0.6333333  0.5579798  0.6758621
#>            Max.
#>    1  1.0000000
#>    2 20.0000000
#>    3  1.0000000
#>    4  0.5000000
#>    5  0.7000000
#>    6  1.0000000
#>    7  1.0000000
#>    8  1.0000000
#>    9  0.4364358
#>   10  0.7200000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1             10             10
#> 

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)
mscurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1     random          1        ROC 0.5746000
#>    2     random          1        PRC 0.5503509
#>    3    poor_er          1        ROC 0.7939000
#>    4    poor_er          1        PRC 0.7728021
#>    5    good_er          1        ROC 0.8392000
#>    6    good_er          1        PRC 0.8704396
#>    7      excel          1        ROC 0.9894000
#>    8      excel          1        PRC 0.9895092
#>    9       perf          1        ROC 1.0000000
#>   10       perf          1        PRC 1.0000000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#> 

## Generate an mspoints object that contains basic evaluation measures
mspoints <- evalmod(mdat, mode = "basic")
mspoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.         Min.     1st Qu.      Median        Mean
#>    1  random  1   rank  0.000000000  0.25000000  0.50000000  0.50000000
#>    2  random  1  score -2.300721014 -0.74953632 -0.15799168 -0.07148314
#>    3  random  1  label -1.000000000 -1.00000000  0.00000000  0.00000000
#>    4  random  1    err  0.425000000  0.45000000  0.46500000  0.46288557
#>    5  random  1    acc  0.495000000  0.52500000  0.53500000  0.53711443
#>    6  random  1     sp  0.000000000  0.30000000  0.56000000  0.53711443
#>    7  random  1     sn  0.000000000  0.28000000  0.56000000  0.53711443
#>    8  random  1   prec  0.000000000  0.52941176  0.54285714  0.54363665
#>    9  random  1    mcc -0.070888121  0.06746320  0.09247931  0.09256534
#>   10  random  1 fscore  0.000000000  0.37333333  0.56038647  0.48753255
#>   11 poor_er  1   rank  0.000000000  0.25000000  0.50000000  0.50000000
#>   12 poor_er  1  score  0.008675501  0.48276124  0.74213209  0.66801432
#>   13 poor_er  1  label -1.000000000 -1.00000000  0.00000000  0.00000000
#>   14 poor_er  1    err  0.275000000  0.30000000  0.33000000  0.35378109
#>   15 poor_er  1    acc  0.500000000  0.60000000  0.67000000  0.64621891
#>   16 poor_er  1     sp  0.000000000  0.43000000  0.71000000  0.64621891
#>   17 poor_er  1     sn  0.000000000  0.40000000  0.71000000  0.64621891
#>   18 poor_er  1   prec  0.500000000  0.62251656  0.71153846  0.71339514
#>   19 poor_er  1    mcc  0.070888121  0.28553298  0.39917840  0.35053171
#>   20 poor_er  1 fscore  0.000000000  0.52631579  0.68817204  0.60130710
#>   21 good_er  1   rank  0.000000000  0.25000000  0.50000000  0.50000000
#>   22 good_er  1  score  0.001998590  0.11605849  0.32419895  0.37498174
#>   23 good_er  1  label -1.000000000 -1.00000000  0.00000000  0.00000000
#>   24 good_er  1    err  0.220000000  0.25500000  0.31000000  0.33124378
#>   25 good_er  1    acc  0.500000000  0.60000000  0.69000000  0.66875622
#>   26 good_er  1     sp  0.000000000  0.40000000  0.77000000  0.66875622
#>   27 good_er  1     sn  0.000000000  0.48000000  0.77000000  0.66875622
#>   28 good_er  1   prec  0.500000000  0.60264901  0.77000000  0.76842171
#>   29 good_er  1    mcc  0.041134503  0.30135369  0.45256964  0.39710164
#>   30 good_er  1 fscore  0.000000000  0.64000000  0.70411985  0.63335434
#>   31   excel  1   rank  0.000000000  0.25000000  0.50000000  0.50000000
#>   32   excel  1  score -2.901311377  0.03298902  1.53323931  1.55713957
#>   33   excel  1  label -1.000000000 -1.00000000  0.00000000  0.00000000
#>   34   excel  1    err  0.050000000  0.13000000  0.25000000  0.25651741
#>   35   excel  1    acc  0.500000000  0.62500000  0.75000000  0.74348259
#>   36   excel  1     sp  0.000000000  0.50000000  0.95000000  0.74348259
#>   37   excel  1     sn  0.000000000  0.50000000  0.95000000  0.74348259
#>   38   excel  1   prec  0.500000000  0.66666667  0.94949495  0.84065254
#>   39   excel  1    mcc  0.070888121  0.38226007  0.57735027  0.56280650
#>   40   excel  1 fscore  0.000000000  0.66666667  0.76335878  0.70508781
#>   41    perf  1   rank  0.000000000  0.25000000  0.50000000  0.50000000
#>   42    perf  1  score  0.000000000  0.00000000  0.50000000  0.50000000
#>   43    perf  1  label -1.000000000 -1.00000000  0.00000000  0.00000000
#>   44    perf  1    err  0.000000000  0.12500000  0.25000000  0.25124378
#>   45    perf  1    acc  0.500000000  0.62500000  0.75000000  0.74875622
#>   46    perf  1     sp  0.000000000  0.50000000  1.00000000  0.74875622
#>   47    perf  1     sn  0.000000000  0.50000000  1.00000000  0.74875622
#>   48    perf  1   prec  0.500000000  0.66666667  1.00000000  0.84609623
#>   49    perf  1    mcc  0.070888121  0.38226007  0.57735027  0.57352524
#>   50    perf  1 fscore  0.000000000  0.66666667  0.76335878  0.71042736
#>        3rd Qu.      Max.
#>    1 0.7500000 1.0000000
#>    2 0.5472949 2.8882527
#>    3 1.0000000 1.0000000
#>    4 0.4750000 0.5050000
#>    5 0.5500000 0.5750000
#>    6 0.7800000 1.0000000
#>    7 0.8000000 1.0000000
#>    8 0.5656566 0.6666667
#>    9 0.1202049 0.1835326
#>   10 0.6425703 0.6827586
#>   11 0.7500000 1.0000000
#>   12 0.8912234 0.9999020
#>   13 1.0000000 1.0000000
#>   14 0.4000000 0.5000000
#>   15 0.7000000 0.7250000
#>   16 0.9000000 1.0000000
#>   17 0.9300000 1.0000000
#>   18 0.7936508 1.0000000
#>   19 0.4238577 0.4586643
#>   20 0.7368421 0.7578125
#>   21 0.7500000 1.0000000
#>   22 0.6060921 0.9826888
#>   23 1.0000000 1.0000000
#>   24 0.4000000 0.5000000
#>   25 0.7450000 0.7800000
#>   26 0.9800000 1.0000000
#>   27 0.9000000 1.0000000
#>   28 0.9555556 1.0000000
#>   29 0.5208859 0.5648911
#>   30 0.7446809 0.7821782
#>   31 0.7500000 1.0000000
#>   32 3.2411655 5.5533171
#>   33 1.0000000 1.0000000
#>   34 0.3750000 0.5000000
#>   35 0.8700000 0.9500000
#>   36 1.0000000 1.0000000
#>   37 1.0000000 1.0000000
#>   38 1.0000000 1.0000000
#>   39 0.7643176 0.9016244
#>   40 0.8715084 0.9509804
#>   41 0.7500000 1.0000000
#>   42 1.0000000 1.0000000
#>   43 1.0000000 1.0000000
#>   44 0.3750000 0.5000000
#>   45 0.8750000 1.0000000
#>   46 1.0000000 1.0000000
#>   47 1.0000000 1.0000000
#>   48 1.0000000 1.0000000
#>   49 0.7745967 1.0000000
#>   50 0.8764045 1.0000000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#> 

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat)
smcurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1    good_er          1        ROC 0.8195000
#>    2    good_er          1        PRC 0.8411248
#>    3    good_er          2        ROC 0.7729000
#>    4    good_er          2        PRC 0.8184637
#>    5    good_er          3        ROC 0.8396000
#>    6    good_er          3        PRC 0.8694958
#>    7    good_er          4        ROC 0.7926000
#>    8    good_er          4        PRC 0.8446957
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1    good_er          1            100            100
#>    2    good_er          2            100            100
#>    3    good_er          3            100            100
#>    4    good_er          4            100            100
#> 

## Generate an smpoints object that contains basic evaluation measures
smpoints <- evalmod(mdat, mode = "basic")
smpoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.          Min.     1st Qu.    Median      Mean   3rd Qu.
#>    1 good_er  1   rank  0.0000000000  0.25000000 0.5000000 0.5000000 0.7500000
#>    2 good_er  1  score  0.0004513370  0.11012505 0.2800428 0.3467666 0.5415409
#>    3 good_er  1  label -1.0000000000 -1.00000000 0.0000000 0.0000000 1.0000000
#>    4 good_er  1    err  0.2500000000  0.27000000 0.3250000 0.3410448 0.4000000
#>    5 good_er  1    acc  0.5000000000  0.60000000 0.6750000 0.6589552 0.7300000
#>    6 good_er  1     sp  0.0000000000  0.41000000 0.7400000 0.6589552 0.9400000
#>    7 good_er  1     sn  0.0000000000  0.44000000 0.7400000 0.6589552 0.9100000
#>    8 good_er  1   prec  0.5000000000  0.60402685 0.7400000 0.7519687 0.8867925
#>    9 good_er  1    mcc  0.0708881205  0.30730550 0.4197919 0.3770060 0.4729429
#>   10 good_er  1 fscore  0.0000000000  0.58666667 0.7003610 0.6213883 0.7319588
#>   11 good_er  2   rank  0.0000000000  0.25000000 0.5000000 0.5000000 0.7500000
#>   12 good_er  2  score  0.0007527706  0.09320446 0.2886411 0.3519440 0.5891042
#>   13 good_er  2  label -1.0000000000 -1.00000000 0.0000000 0.0000000 1.0000000
#>   14 good_er  2    err  0.2600000000  0.30000000 0.3550000 0.3642289 0.4150000
#>   15 good_er  2    acc  0.4950000000  0.58500000 0.6450000 0.6357711 0.7000000
#>   16 good_er  2     sp  0.0000000000  0.35000000 0.7000000 0.6357711 0.9500000
#>   17 good_er  2     sn  0.0000000000  0.45000000 0.7000000 0.6357711 0.8500000
#>   18 good_er  2   prec  0.4974874372  0.56862745 0.7000000 0.7342268 0.9000000
#>   19 good_er  2    mcc -0.0708881205  0.23860996 0.3470110 0.3223944 0.4222570
#>   20 good_er  2 fscore  0.0000000000  0.60000000 0.6783626 0.6013090 0.6926407
#>   21 good_er  3   rank  0.0000000000  0.25000000 0.5000000 0.5000000 0.7500000
#>   22 good_er  3  score  0.0001600914  0.12843849 0.3171778 0.3767497 0.6019583
#>   23 good_er  3  label -1.0000000000 -1.00000000 0.0000000 0.0000000 1.0000000
#>   24 good_er  3    err  0.2300000000  0.25500000 0.3000000 0.3310448 0.4100000
#>   25 good_er  3    acc  0.5000000000  0.59000000 0.7000000 0.6689552 0.7450000
#>   26 good_er  3     sp  0.0000000000  0.42000000 0.7500000 0.6689552 0.9800000
#>   27 good_er  3     sn  0.0000000000  0.48000000 0.7500000 0.6689552 0.9200000
#>   28 good_er  3   prec  0.5000000000  0.61333333 0.7500000 0.7680459 0.9591837
#>   29 good_er  3    mcc  0.0000000000  0.28553298 0.4685213 0.3960926 0.5123179
#>   30 good_er  3 fscore  0.0000000000  0.64000000 0.6940299 0.6333007 0.7475728
#>   31 good_er  4   rank  0.0000000000  0.25000000 0.5000000 0.5000000 0.7500000
#>   32 good_er  4  score  0.0079090710  0.10495864 0.2272597 0.3248899 0.5033557
#>   33 good_er  4  label -1.0000000000 -1.00000000 0.0000000 0.0000000 1.0000000
#>   34 good_er  4    err  0.2450000000  0.27500000 0.3600000 0.3544279 0.4050000
#>   35 good_er  4    acc  0.5000000000  0.59500000 0.6400000 0.6455721 0.7250000
#>   36 good_er  4     sp  0.0000000000  0.38000000 0.7200000 0.6455721 0.9800000
#>   37 good_er  4     sn  0.0000000000  0.48000000 0.7200000 0.6455721 0.8800000
#>   38 good_er  4   prec  0.5000000000  0.58940397 0.7200000 0.7493018 0.9600000
#>   39 good_er  4    mcc  0.0000000000  0.26561731 0.3241764 0.3452893 0.4802181
#>   40 good_er  4 fscore  0.0000000000  0.63576159 0.6837607 0.6124705 0.7020408
#>           Max.
#>    1 1.0000000
#>    2 0.9882066
#>    3 1.0000000
#>    4 0.5000000
#>    5 0.7500000
#>    6 1.0000000
#>    7 1.0000000
#>    8 1.0000000
#>    9 0.5083023
#>   10 0.7644444
#>   11 1.0000000
#>   12 0.9992423
#>   13 1.0000000
#>   14 0.5050000
#>   15 0.7400000
#>   16 1.0000000
#>   17 1.0000000
#>   18 1.0000000
#>   19 0.5205859
#>   20 0.7150259
#>   21 1.0000000
#>   22 0.9811281
#>   23 1.0000000
#>   24 0.5000000
#>   25 0.7700000
#>   26 1.0000000
#>   27 1.0000000
#>   28 1.0000000
#>   29 0.5509496
#>   30 0.7870370
#>   31 1.0000000
#>   32 0.9706821
#>   33 1.0000000
#>   34 0.5000000
#>   35 0.7550000
#>   36 1.0000000
#>   37 1.0000000
#>   38 1.0000000
#>   39 0.5465357
#>   40 0.7272727
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1    good_er          1            100            100
#>    2    good_er          2            100            100
#>    3    good_er          3            100            100
#>    4    good_er          4            100            100
#> 

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mmcurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat)
mmcurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1     random          1        ROC 0.4621000
#>    2     random          1        PRC 0.4961154
#>    3    poor_er          1        ROC 0.7977000
#>    4    poor_er          1        PRC 0.7492474
#>    5    good_er          1        ROC 0.8350000
#>    6    good_er          1        PRC 0.8629297
#>    7      excel          1        ROC 0.9767000
#>    8      excel          1        PRC 0.9802465
#>    9       perf          1        ROC 1.0000000
#>   10       perf          1        PRC 1.0000000
#>   11     random          2        ROC 0.4878000
#>   12     random          2        PRC 0.4914722
#>   13    poor_er          2        ROC 0.7951000
#>   14    poor_er          2        PRC 0.7154885
#>   15    good_er          2        ROC 0.7460000
#>   16    good_er          2        PRC 0.7733400
#>   17      excel          2        ROC 0.9800000
#>   18      excel          2        PRC 0.9819113
#>   19       perf          2        ROC 1.0000000
#>   20       perf          2        PRC 1.0000000
#>   21     random          3        ROC 0.4717000
#>   22     random          3        PRC 0.4919954
#>   23    poor_er          3        ROC 0.8220000
#>   24    poor_er          3        PRC 0.7579889
#>   25    good_er          3        ROC 0.7921000
#>   26    good_er          3        PRC 0.8286359
#>   27      excel          3        ROC 0.9898000
#>   28      excel          3        PRC 0.9898761
#>   29       perf          3        ROC 1.0000000
#>   30       perf          3        PRC 1.0000000
#>   31     random          4        ROC 0.4689000
#>   32     random          4        PRC 0.4884887
#>   33    poor_er          4        ROC 0.8161000
#>   34    poor_er          4        PRC 0.7402335
#>   35    good_er          4        ROC 0.8187000
#>   36    good_er          4        PRC 0.8607595
#>   37      excel          4        ROC 0.9835000
#>   38      excel          4        PRC 0.9853430
#>   39       perf          4        ROC 1.0000000
#>   40       perf          4        PRC 1.0000000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#>    6     random          2            100            100
#>    7    poor_er          2            100            100
#>    8    good_er          2            100            100
#>    9      excel          2            100            100
#>   10       perf          2            100            100
#>   11     random          3            100            100
#>   12    poor_er          3            100            100
#>   13    good_er          3            100            100
#>   14      excel          3            100            100
#>   15       perf          3            100            100
#>   16     random          4            100            100
#>   17    poor_er          4            100            100
#>   18    good_er          4            100            100
#>   19      excel          4            100            100
#>   20       perf          4            100            100
#> 

## Generate an mmpoints object that contains basic evaluation measures
mmpoints <- evalmod(mdat, mode = "basic")
mmpoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.          Min.     1st Qu.       Median        Mean
#>    1  random  1   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>    2  random  1  score -3.0207576069 -0.98492756 -0.086158764 -0.20169748
#>    3  random  1  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>    4  random  1    err  0.4750000000  0.50000000  0.525000000  0.51885572
#>    5  random  1    acc  0.4450000000  0.46500000  0.475000000  0.48114428
#>    6  random  1     sp  0.0000000000  0.20000000  0.450000000  0.48114428
#>    7  random  1     sn  0.0000000000  0.24000000  0.450000000  0.48114428
#>    8  random  1   prec  0.3333333333  0.46835443  0.478632479  0.49483986
#>    9  random  1    mcc -0.1297173419 -0.08284725 -0.051140359 -0.03942501
#>   10  random  1 fscore  0.0000000000  0.32432432  0.452261307  0.43406167
#>   11 poor_er  1   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   12 poor_er  1  score  0.0058608269  0.49481298  0.751698795  0.66454160
#>   13 poor_er  1  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   14 poor_er  1    err  0.2400000000  0.30000000  0.335000000  0.35189055
#>   15 poor_er  1    acc  0.5000000000  0.58500000  0.665000000  0.64810945
#>   16 poor_er  1     sp  0.0000000000  0.48000000  0.700000000  0.64810945
#>   17 poor_er  1     sn  0.0000000000  0.39000000  0.700000000  0.64810945
#>   18 poor_er  1   prec  0.5000000000  0.65333333  0.700000000  0.70047878
#>   19 poor_er  1    mcc  0.0708881205  0.27238140  0.365977656  0.35361698
#>   20 poor_er  1 fscore  0.0000000000  0.52000000  0.684931507  0.59970617
#>   21 good_er  1   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   22 good_er  1  score  0.0010918662  0.10790625  0.280723667  0.36221025
#>   23 good_er  1  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   24 good_er  1    err  0.2000000000  0.25500000  0.320000000  0.33333333
#>   25 good_er  1    acc  0.5000000000  0.60000000  0.680000000  0.66666667
#>   26 good_er  1     sp  0.0000000000  0.39000000  0.760000000  0.66666667
#>   27 good_er  1     sn  0.0000000000  0.45000000  0.760000000  0.66666667
#>   28 good_er  1   prec  0.5000000000  0.59333333  0.760000000  0.76428454
#>   29 good_er  1    mcc  0.0708881205  0.29247189  0.411813845  0.39561088
#>   30 good_er  1 fscore  0.0000000000  0.60000000  0.706319703  0.63077171
#>   31   excel  1   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   32   excel  1  score -2.7202907052 -0.02580676  1.501440337  1.49985492
#>   33   excel  1  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   34   excel  1    err  0.0550000000  0.14000000  0.255000000  0.26283582
#>   35   excel  1    acc  0.5000000000  0.62000000  0.745000000  0.73716418
#>   36   excel  1     sp  0.0000000000  0.49000000  0.940000000  0.73716418
#>   37   excel  1     sn  0.0000000000  0.50000000  0.940000000  0.73716418
#>   38   excel  1   prec  0.5000000000  0.66000000  0.940000000  0.83512742
#>   39   excel  1    mcc  0.0708881205  0.36753360  0.569651921  0.54871513
#>   40   excel  1 fscore  0.0000000000  0.66666667  0.753086420  0.69931155
#>   41    perf  1   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   42    perf  1  score  0.0000000000  0.00000000  0.500000000  0.50000000
#>   43    perf  1  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   44    perf  1    err  0.0000000000  0.12500000  0.250000000  0.25124378
#>   45    perf  1    acc  0.5000000000  0.62500000  0.750000000  0.74875622
#>   46    perf  1     sp  0.0000000000  0.50000000  1.000000000  0.74875622
#>   47    perf  1     sn  0.0000000000  0.50000000  1.000000000  0.74875622
#>   48    perf  1   prec  0.5000000000  0.66666667  1.000000000  0.84609623
#>   49    perf  1    mcc  0.0708881205  0.38226007  0.577350269  0.57352524
#>   50    perf  1 fscore  0.0000000000  0.66666667  0.763358779  0.71042736
#>   51  random  2   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   52  random  2  score -2.6316929933 -0.64398968 -0.112485604 -0.01881380
#>   53  random  2  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   54  random  2    err  0.4750000000  0.49000000  0.500000000  0.50606965
#>   55  random  2    acc  0.4350000000  0.49000000  0.500000000  0.49393035
#>   56  random  2     sp  0.0000000000  0.26000000  0.470000000  0.49393035
#>   57  random  2     sn  0.0000000000  0.25000000  0.470000000  0.49393035
#>   58  random  2   prec  0.0000000000  0.46969697  0.500000000  0.48652140
#>   59  random  2    mcc -0.1319202225 -0.04804344  0.000000000 -0.01137431
#>   60  random  2 fscore  0.0000000000  0.32894737  0.463054187  0.44359450
#>   61 poor_er  2   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   62 poor_er  2  score  0.0106671425  0.47649621  0.756978310  0.66284643
#>   63 poor_er  2  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   64 poor_er  2    err  0.2100000000  0.27500000  0.360000000  0.35318408
#>   65 poor_er  2    acc  0.5000000000  0.57000000  0.640000000  0.64681592
#>   66 poor_er  2     sp  0.0000000000  0.48000000  0.740000000  0.64681592
#>   67 poor_er  2     sn  0.0000000000  0.35000000  0.740000000  0.64681592
#>   68 poor_er  2   prec  0.5000000000  0.63636364  0.692307692  0.67780698
#>   69 poor_er  2    mcc  0.0219317232  0.20201247  0.351561525  0.34436601
#>   70 poor_er  2 fscore  0.0000000000  0.46666667  0.689655172  0.59372053
#>   71 good_er  2   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   72 good_er  2  score  0.0025241743  0.09296394  0.234898708  0.30399918
#>   73 good_er  2  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   74 good_er  2    err  0.3050000000  0.33500000  0.355000000  0.37761194
#>   75 good_er  2    acc  0.4950000000  0.59000000  0.645000000  0.62238806
#>   76 good_er  2     sp  0.0000000000  0.36000000  0.670000000  0.62238806
#>   77 good_er  2     sn  0.0000000000  0.42000000  0.670000000  0.62238806
#>   78 good_er  2   prec  0.4974874372  0.57692308  0.663157895  0.70784945
#>   79 good_er  2    mcc -0.0708881205  0.26252534  0.314938446  0.29466351
#>   80 good_er  2 fscore  0.0000000000  0.56000000  0.662207358  0.58363614
#>   81   excel  2   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   82   excel  2  score -3.2487982202 -0.03931856  1.271241656  1.44090081
#>   83   excel  2  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   84   excel  2    err  0.0700000000  0.13000000  0.250000000  0.26119403
#>   85   excel  2    acc  0.5000000000  0.62500000  0.750000000  0.73880597
#>   86   excel  2     sp  0.0000000000  0.50000000  0.920000000  0.73880597
#>   87   excel  2     sn  0.0000000000  0.50000000  0.920000000  0.73880597
#>   88   excel  2   prec  0.5000000000  0.66666667  0.920000000  0.83637640
#>   89   excel  2    mcc  0.0708881205  0.38226007  0.577350269  0.55328872
#>   90   excel  2 fscore  0.0000000000  0.66666667  0.763358779  0.70062050
#>   91    perf  2   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>   92    perf  2  score  0.0000000000  0.00000000  0.500000000  0.50000000
#>   93    perf  2  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>   94    perf  2    err  0.0000000000  0.12500000  0.250000000  0.25124378
#>   95    perf  2    acc  0.5000000000  0.62500000  0.750000000  0.74875622
#>   96    perf  2     sp  0.0000000000  0.50000000  1.000000000  0.74875622
#>   97    perf  2     sn  0.0000000000  0.50000000  1.000000000  0.74875622
#>   98    perf  2   prec  0.5000000000  0.66666667  1.000000000  0.84609623
#>   99    perf  2    mcc  0.0708881205  0.38226007  0.577350269  0.57352524
#>  100    perf  2 fscore  0.0000000000  0.66666667  0.763358779  0.71042736
#>  101  random  3   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>  102  random  3  score -2.9804105202 -0.64666644 -0.003867533  0.06581216
#>  103  random  3  label -1.0000000000 -1.00000000  0.000000000  0.00000000
#>  104  random  3    err  0.4850000000  0.50000000  0.515000000  0.51407960
#>  105  random  3    acc  0.4550000000  0.47500000  0.485000000  0.48592040
#>  106  random  3     sp  0.0000000000  0.21000000  0.480000000  0.48592040
#>  107  random  3     sn  0.0000000000  0.23000000  0.480000000  0.48592040
#>  108  random  3   prec  0.3333333333  0.47619048  0.485507246  0.49132990
#>  109  random  3    mcc -0.1019637753 -0.06079653 -0.032025631 -0.03290522
#>  110  random  3 fscore  0.0000000000  0.31578947  0.475247525  0.43723044
#>  111 poor_er  3   rank  0.0000000000  0.25000000  0.500000000  0.50000000
#>           3rd Qu.       Max.
#>    1  0.750000000 1.00000000
#>    2  0.509261760 2.85400950
#>    3  1.000000000 1.00000000
#>    4  0.535000000 0.55500000
#>    5  0.500000000 0.52500000
#>    6  0.740000000 1.00000000
#>    7  0.700000000 1.00000000
#>    8  0.500000000 1.00000000
#>    9  0.000000000 0.10050378
#>   10  0.560000000 0.67114094
#>   11  0.750000000 1.00000000
#>   12  0.897220281 0.99523044
#>   13  1.000000000 1.00000000
#>   14  0.415000000 0.50000000
#>   15  0.700000000 0.76000000
#>   16  0.890000000 1.00000000
#>   17  0.980000000 1.00000000
#>   18  0.766666667 1.00000000
#>   19  0.440481692 0.57298754
#>   20  0.743494424 0.80165289
#>   21  0.750000000 1.00000000
#>   22  0.573394627 0.97607704
#>   23  1.000000000 1.00000000
#>   24  0.400000000 0.50000000
#>   25  0.745000000 0.80000000
#>   26  0.950000000 1.00000000
#>   27  0.890000000 1.00000000
#>   28  0.914893617 1.00000000
#>   29  0.510127578 0.61237244
#>   30  0.742081448 0.77777778
#>   31  0.750000000 1.00000000
#>   32  3.007937414 7.01515299
#>   33  1.000000000 1.00000000
#>   34  0.380000000 0.50000000
#>   35  0.860000000 0.94500000
#>   36  1.000000000 1.00000000
#>   37  0.990000000 1.00000000
#>   38  1.000000000 1.00000000
#>   39  0.743660399 0.89004450
#>   40  0.859649123 0.94472362
#>   41  0.750000000 1.00000000
#>   42  1.000000000 1.00000000
#>   43  1.000000000 1.00000000
#>   44  0.375000000 0.50000000
#>   45  0.875000000 1.00000000
#>   46  1.000000000 1.00000000
#>   47  1.000000000 1.00000000
#>   48  1.000000000 1.00000000
#>   49  0.774596669 1.00000000
#>   50  0.876404494 1.00000000
#>   51  0.750000000 1.00000000
#>   52  0.732228468 3.03078470
#>   53  1.000000000 1.00000000
#>   54  0.510000000 0.56500000
#>   55  0.510000000 0.52500000
#>   56  0.750000000 1.00000000
#>   57  0.760000000 1.00000000
#>   58  0.508196721 0.61904762
#>   59  0.024559577 0.08161943
#>   60  0.609375000 0.66894198
#>   61  0.750000000 1.00000000
#>   62  0.891929040 0.99907303
#>   63  1.000000000 1.00000000
#>   64  0.430000000 0.50000000
#>   65  0.725000000 0.79000000
#>   66  0.850000000 1.00000000
#>   67  0.980000000 1.00000000
#>   68  0.725000000 1.00000000
#>   69  0.507010830 0.58576894
#>   70  0.770428016 0.80686695
#>   71  0.750000000 1.00000000
#>   72  0.481955469 0.98203554
#>   73  1.000000000 1.00000000
#>   74  0.410000000 0.50500000
#>   75  0.665000000 0.69500000
#>   76  0.920000000 1.00000000
#>   77  0.860000000 1.00000000
#>   78  0.843137255 1.00000000
#>   79  0.341712857 0.41314394
#>   80  0.690909091 0.71244635
#>   81  0.750000000 1.00000000
#>   82  3.074395095 5.31788333
#>   83  1.000000000 1.00000000
#>   84  0.375000000 0.50000000
#>   85  0.870000000 0.93000000
#>   86  1.000000000 1.00000000
#>   87  1.000000000 1.00000000
#>   88  1.000000000 1.00000000
#>   89  0.763097317 0.86276528
#>   90  0.870056497 0.92708333
#>   91  0.750000000 1.00000000
#>   92  1.000000000 1.00000000
#>   93  1.000000000 1.00000000
#>   94  0.375000000 0.50000000
#>   95  0.875000000 1.00000000
#>   96  1.000000000 1.00000000
#>   97  1.000000000 1.00000000
#>   98  1.000000000 1.00000000
#>   99  0.774596669 1.00000000
#>  100  0.876404494 1.00000000
#>  101  0.750000000 1.00000000
#>  102  0.815393972 2.75725126
#>  103  1.000000000 1.00000000
#>  104  0.525000000 0.54500000
#>  105  0.500000000 0.51500000
#>  106  0.730000000 1.00000000
#>  107  0.710000000 1.00000000
#>  108  0.500000000 1.00000000
#>  109 -0.005073855 0.07088812
#>  110  0.570281124 0.66666667
#>  111  0.750000000 1.00000000
#>  [ reached getOption("max.print") -- omitted 89 rows ]
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#>    6     random          2            100            100
#>    7    poor_er          2            100            100
#>    8    good_er          2            100            100
#>    9      excel          2            100            100
#>   10       perf          2            100            100
#>   11     random          3            100            100
#>   12    poor_er          3            100            100
#>   13    good_er          3            100            100
#>   14      excel          3            100            100
#>   15       perf          3            100            100
#>   16     random          4            100            100
#>   17    poor_er          4            100            100
#>   18    good_er          4            100            100
#>   19      excel          4            100            100
#>   20       perf          4            100            100
#> 

##################################################
### N-fold cross validation datasets
###

## Load test data
data(M2N50F5)

## Speficy nessesary columns to create mdat
cvdat <- mmdata(nfold_df = M2N50F5, score_cols = c(1, 2),
                lab_col = 3, fold_col = 4,
                modnames = c("m1", "m2"), dsids = 1:5)

## Generate an mmcurve object that contains ROC and Precision-Recall curves
cvcurves <- evalmod(cvdat)
cvcurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1         m1          1        ROC 1.0000000
#>    2         m1          1        PRC 1.0000000
#>    3         m1          2        ROC 0.4166667
#>    4         m1          2        PRC 0.5164199
#>    5         m1          3        ROC 0.2000000
#>    6         m1          3        PRC 0.4891743
#>    7         m1          4        ROC 0.7916667
#>    8         m1          4        PRC 0.7728152
#>    9         m1          5        ROC 0.4400000
#>   10         m1          5        PRC 0.4266312
#>   11         m2          1        ROC 0.4000000
#>   12         m2          1        PRC 0.4247188
#>   13         m2          2        ROC 0.7083333
#>   14         m2          2        PRC 0.6568625
#>   15         m2          3        ROC 0.8400000
#>   16         m2          3        PRC 0.9057736
#>   17         m2          4        ROC 0.7916667
#>   18         m2          4        PRC 0.8527712
#>   19         m2          5        ROC 0.4000000
#>   20         m2          5        PRC 0.4247188
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1              5              5
#>    2         m1          2              4              6
#>    3         m1          3              5              5
#>    4         m1          4              6              4
#>    5         m1          5              5              5
#>    6         m2          1              5              5
#>    7         m2          2              4              6
#>    8         m2          3              5              5
#>    9         m2          4              6              4
#>   10         m2          5              5              5
#> 

## Generate an mmpoints object that contains basic evaluation measures
cvpoints <- evalmod(cvdat, mode = "basic")
cvpoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>      Model ID  Meas.       Min.     1st Qu.       Median         Mean
#>    1    m1  1   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>    2    m1  1  score -1.5761733 -0.92376396 -0.002327284  0.113500523
#>    3    m1  1  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>    4    m1  1    err  0.0000000  0.15000000  0.300000000  0.272727273
#>    5    m1  1    acc  0.5000000  0.60000000  0.700000000  0.727272727
#>    6    m1  1     sp  0.0000000  0.50000000  1.000000000  0.727272727
#>    7    m1  1     sn  0.0000000  0.50000000  1.000000000  0.727272727
#>    8    m1  1   prec  0.5000000  0.66964286  1.000000000  0.838924964
#>    9    m1  1    mcc  0.3333333  0.50000000  0.654653671  0.623218574
#>   10    m1  1 fscore  0.0000000  0.61904762  0.750000000  0.676023471
#>   11    m1  2   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   12    m1  2  score -1.8425134 -1.08765052 -0.450813882 -0.375652956
#>   13    m1  2  label -1.0000000 -1.00000000  1.000000000  0.200000000
#>   14    m1  2    err  0.4000000  0.45000000  0.500000000  0.536363636
#>   15    m1  2    acc  0.3000000  0.40000000  0.500000000  0.463636364
#>   16    m1  2     sp  0.0000000  0.25000000  0.500000000  0.454545455
#>   17    m1  2     sn  0.0000000  0.16666667  0.500000000  0.469696970
#>   18    m1  2   prec  0.0000000  0.41666667  0.555555556  0.450180375
#>   19    m1  2    mcc -0.4082483 -0.27216553 -0.102062073 -0.125094358
#>   20    m1  2 fscore  0.0000000  0.23611111  0.545454545  0.439152766
#>   21    m1  3   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   22    m1  3  score -1.1558752 -0.16251342  0.389025335  0.414373732
#>   23    m1  3  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>   24    m1  3    err  0.4000000  0.50000000  0.600000000  0.636363636
#>   25    m1  3    acc  0.1000000  0.25000000  0.400000000  0.363636364
#>   26    m1  3     sp  0.0000000  0.00000000  0.200000000  0.363636364
#>   27    m1  3     sn  0.0000000  0.20000000  0.200000000  0.363636364
#>   28    m1  3   prec  0.1666667  0.26785714  0.375000000  0.459559885
#>   29    m1  3    mcc -0.8164966 -0.60000000 -0.408248290 -0.355290715
#>   30    m1  3 fscore  0.0000000  0.21111111  0.285714286  0.318732278
#>   31    m1  4   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   32    m1  4  score -1.5462252 -1.19762686 -0.400691531 -0.141998509
#>   33    m1  4  label -1.0000000 -1.00000000 -1.000000000 -0.200000000
#>   34    m1  4    err  0.2000000  0.30000000  0.400000000  0.372727273
#>   35    m1  4    acc  0.4000000  0.60000000  0.600000000  0.627272727
#>   36    m1  4     sp  0.0000000  0.41666667  0.666666667  0.606060606
#>   37    m1  4     sn  0.0000000  0.50000000  0.750000000  0.659090909
#>   38    m1  4   prec  0.4000000  0.50000000  0.571428571  0.652958153
#>   39    m1  4    mcc  0.1666667  0.27216553  0.408248290  0.379646701
#>   40    m1  4 fscore  0.0000000  0.53571429  0.600000000  0.544137681
#>   41    m1  5   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   42    m1  5  score -1.5944199 -0.50932491 -0.113562820  0.006988569
#>   43    m1  5  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>   44    m1  5    err  0.4000000  0.50000000  0.500000000  0.527272727
#>   45    m1  5    acc  0.3000000  0.40000000  0.500000000  0.472727273
#>   46    m1  5     sp  0.0000000  0.30000000  0.400000000  0.472727273
#>   47    m1  5     sn  0.0000000  0.10000000  0.400000000  0.472727273
#>   48    m1  5   prec  0.0000000  0.16666667  0.500000000  0.350937951
#>   49    m1  5    mcc -0.5000000 -0.21821789  0.000000000 -0.077777778
#>   50    m1  5 fscore  0.0000000  0.12500000  0.444444444  0.391172968
#>   51    m2  1   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   52    m2  1  score -1.5666375 -0.93714170 -0.338617003 -0.081054576
#>   53    m2  1  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>   54    m2  1    err  0.5000000  0.50000000  0.500000000  0.545454545
#>   55    m2  1    acc  0.4000000  0.40000000  0.500000000  0.454545455
#>   56    m2  1     sp  0.0000000  0.20000000  0.400000000  0.454545455
#>   57    m2  1     sn  0.0000000  0.20000000  0.400000000  0.454545455
#>   58    m2  1   prec  0.0000000  0.36666667  0.444444444  0.373304473
#>   59    m2  1    mcc -0.3333333 -0.21821789 -0.200000000 -0.144789161
#>   60    m2  1 fscore  0.0000000  0.26785714  0.444444444  0.389008466
#>   61    m2  2   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   62    m2  2  score -0.9422955 -0.51338997  0.168580899 -0.003990330
#>   63    m2  2  label -1.0000000 -1.00000000  1.000000000  0.200000000
#>   64    m2  2    err  0.2000000  0.30000000  0.400000000  0.409090909
#>   65    m2  2    acc  0.3000000  0.45000000  0.600000000  0.590909091
#>   66    m2  2     sp  0.0000000  0.50000000  0.750000000  0.613636364
#>   67    m2  2     sn  0.0000000  0.25000000  0.666666667  0.575757576
#>   68    m2  2   prec  0.0000000  0.55000000  0.666666667  0.570995671
#>   69    m2  2    mcc -0.4082483  0.08908708  0.356348323  0.244147488
#>   70    m2  2 fscore  0.0000000  0.34722222  0.727272727  0.548311285
#>   71    m2  3   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   72    m2  3  score -1.0962763 -0.40057827 -0.167608702  0.082174350
#>   73    m2  3  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>   74    m2  3    err  0.1000000  0.25000000  0.400000000  0.345454545
#>   75    m2  3    acc  0.5000000  0.55000000  0.600000000  0.654545455
#>   76    m2  3     sp  0.0000000  0.30000000  0.800000000  0.654545455
#>   77    m2  3     sn  0.0000000  0.50000000  0.800000000  0.654545455
#>   78    m2  3   prec  0.5000000  0.56349206  0.800000000  0.781240981
#>   79    m2  3    mcc  0.0000000  0.33333333  0.408248290  0.429364789
#>   80    m2  3 fscore  0.0000000  0.59340659  0.666666667  0.612175199
#>   81    m2  4   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   82    m2  4  score -1.1654721 -0.35410360 -0.099106516  0.103674375
#>   83    m2  4  label -1.0000000 -1.00000000 -1.000000000 -0.200000000
#>   84    m2  4    err  0.1000000  0.25000000  0.400000000  0.372727273
#>   85    m2  4    acc  0.4000000  0.50000000  0.600000000  0.627272727
#>   86    m2  4     sp  0.0000000  0.25000000  0.666666667  0.606060606
#>   87    m2  4     sn  0.0000000  0.62500000  0.750000000  0.659090909
#>   88    m2  4   prec  0.3750000  0.43650794  0.600000000  0.681637807
#>   89    m2  4    mcc -0.1020621  0.25000000  0.408248290  0.369241846
#>   90    m2  4 fscore  0.0000000  0.52272727  0.600000000  0.561158538
#>   91    m2  5   rank  0.0000000  0.25000000  0.500000000  0.500000000
#>   92    m2  5  score -1.7908147 -1.14209817  0.032207934 -0.329061088
#>   93    m2  5  label -1.0000000 -1.00000000  0.000000000  0.000000000
#>   94    m2  5    err  0.5000000  0.50000000  0.500000000  0.545454545
#>   95    m2  5    acc  0.4000000  0.40000000  0.500000000  0.454545455
#>   96    m2  5     sp  0.0000000  0.20000000  0.400000000  0.454545455
#>   97    m2  5     sn  0.0000000  0.20000000  0.400000000  0.454545455
#>   98    m2  5   prec  0.0000000  0.36666667  0.444444444  0.373304473
#>   99    m2  5    mcc -0.3333333 -0.21821789 -0.200000000 -0.144789161
#>  100    m2  5 fscore  0.0000000  0.26785714  0.444444444  0.389008466
#>          3rd Qu.      Max.
#>    1  0.75000000 1.0000000
#>    2  1.06708368 2.0606025
#>    3  1.00000000 1.0000000
#>    4  0.40000000 0.5000000
#>    5  0.85000000 1.0000000
#>    6  1.00000000 1.0000000
#>    7  1.00000000 1.0000000
#>    8  1.00000000 1.0000000
#>    9  0.81649658 1.0000000
#>   10  0.86111111 1.0000000
#>   11  0.75000000 1.0000000
#>   12 -0.08830171 2.2002973
#>   13  1.00000000 1.0000000
#>   14  0.60000000 0.7000000
#>   15  0.55000000 0.6000000
#>   16  0.62500000 1.0000000
#>   17  0.75000000 1.0000000
#>   18  0.60000000 0.6666667
#>   19  0.00000000 0.1666667
#>   20  0.66666667 0.7500000
#>   21  0.75000000 1.0000000
#>   22  1.01753386 1.6870336
#>   23  1.00000000 1.0000000
#>   24  0.75000000 0.9000000
#>   25  0.50000000 0.6000000
#>   26  0.70000000 1.0000000
#>   27  0.50000000 1.0000000
#>   28  0.50000000 1.0000000
#>   29 -0.21821789 0.3333333
#>   30  0.39743590 0.6666667
#>   31  0.75000000 1.0000000
#>   32  1.00867637 1.5387145
#>   33  1.00000000 1.0000000
#>   34  0.40000000 0.6000000
#>   35  0.70000000 0.8000000
#>   36  0.91666667 1.0000000
#>   37  1.00000000 1.0000000
#>   38  0.83333333 1.0000000
#>   39  0.40824829 0.6123724
#>   40  0.66666667 0.7272727
#>   41  0.75000000 1.0000000
#>   42  0.45270677 1.9237289
#>   43  1.00000000 1.0000000
#>   44  0.60000000 0.7000000
#>   45  0.50000000 0.6000000
#>   46  0.60000000 1.0000000
#>   47  0.80000000 1.0000000
#>   48  0.50000000 0.5714286
#>   49  0.00000000 0.3333333
#>   50  0.64102564 0.7142857
#>   51  0.75000000 1.0000000
#>   52  0.43266892 2.8510651
#>   53  1.00000000 1.0000000
#>   54  0.60000000 0.6000000
#>   55  0.50000000 0.5000000
#>   56  0.70000000 1.0000000
#>   57  0.70000000 1.0000000
#>   58  0.50000000 0.5000000
#>   59  0.00000000 0.0000000
#>   60  0.55844156 0.6666667
#>   61  0.75000000 1.0000000
#>   62  0.39797893 0.7890534
#>   63  1.00000000 1.0000000
#>   64  0.55000000 0.7000000
#>   65  0.70000000 0.8000000
#>   66  0.75000000 1.0000000
#>   67  0.91666667 1.0000000
#>   68  0.75000000 0.8333333
#>   69  0.40824829 0.6123724
#>   70  0.78461538 0.8571429
#>   71  0.75000000 1.0000000
#>   72  0.35464019 2.0647833
#>   73  1.00000000 1.0000000
#>   74  0.45000000 0.5000000
#>   75  0.75000000 0.9000000
#>   76  1.00000000 1.0000000
#>   77  0.80000000 1.0000000
#>   78  1.00000000 1.0000000
#>   79  0.60000000 0.8164966
#>   80  0.73863636 0.8888889
#>   81  0.75000000 1.0000000
#>   82  0.79072036 1.4431452
#>   83  1.00000000 1.0000000
#>   84  0.50000000 0.6000000
#>   85  0.75000000 0.9000000
#>   86  1.00000000 1.0000000
#>   87  0.75000000 1.0000000
#>   88  1.00000000 1.0000000
#>   89  0.58333333 0.8017837
#>   90  0.66666667 0.8571429
#>   91  0.75000000 1.0000000
#>   92  0.31323748 0.7002948
#>   93  1.00000000 1.0000000
#>   94  0.60000000 0.6000000
#>   95  0.50000000 0.5000000
#>   96  0.70000000 1.0000000
#>   97  0.70000000 1.0000000
#>   98  0.50000000 0.5000000
#>   99  0.00000000 0.0000000
#>  100  0.55844156 0.6666667
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1              5              5
#>    2         m1          2              4              6
#>    3         m1          3              5              5
#>    4         m1          4              6              4
#>    5         m1          5              5              5
#>    6         m2          1              5              5
#>    7         m2          2              4              6
#>    8         m2          3              5              5
#>    9         m2          4              6              4
#>   10         m2          5              5              5
#> 

## Specify mmdata arguments from evalmod
cvcurves2 <- evalmod(nfold_df = M2N50F5, score_cols = c(1, 2),
                     lab_col = 3, fold_col = 4,
                     modnames = c("m1", "m2"), dsids = 1:5)
cvcurves2

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1         m1          1        ROC 1.0000000
#>    2         m1          1        PRC 1.0000000
#>    3         m1          2        ROC 0.4166667
#>    4         m1          2        PRC 0.5164199
#>    5         m1          3        ROC 0.2000000
#>    6         m1          3        PRC 0.4891743
#>    7         m1          4        ROC 0.7916667
#>    8         m1          4        PRC 0.7728152
#>    9         m1          5        ROC 0.4400000
#>   10         m1          5        PRC 0.4266312
#>   11         m2          1        ROC 0.4000000
#>   12         m2          1        PRC 0.4247188
#>   13         m2          2        ROC 0.7083333
#>   14         m2          2        PRC 0.6568625
#>   15         m2          3        ROC 0.8400000
#>   16         m2          3        PRC 0.9057736
#>   17         m2          4        ROC 0.7916667
#>   18         m2          4        PRC 0.8527712
#>   19         m2          5        ROC 0.4000000
#>   20         m2          5        PRC 0.4247188
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1              5              5
#>    2         m1          2              4              6
#>    3         m1          3              5              5
#>    4         m1          4              6              4
#>    5         m1          5              5              5
#>    6         m2          1              5              5
#>    7         m2          2              4              6
#>    8         m2          3              5              5
#>    9         m2          4              6              4
#>   10         m2          5              5              5
#> 

##################################################
### AUC with the U statistic
###

## mode = "aucroc" returns 'aucroc' S3 object
data(P10N10)

# 'aucroc' S3 object
uauc1 <- evalmod(scores = P10N10$scores, labels = P10N10$labels,
                 mode="aucroc")

# print 'aucroc'
uauc1

#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of negatives # of positives
#>    1         m1          1             10             10
#> 
#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID  AUC  U
#>    1         m1          1 0.72 72
#> 

# as.data.frame 'aucroc'
as.data.frame(uauc1)

#>   modnames dsids aucs ustats
#> 1       m1     1 0.72     72

## It is 2-3 times faster than mode = "rocprc"
# A sample of 100,000
samp1 <- create_sim_samples(1, 50000, 50000)

# a function to test mode = "rocprc"
func_evalmod_rocprc <- function(samp) {
   curves <- evalmod(scores = samp$scores, labels = samp$labels)
   aucs <- auc(curves)
}

# a function to test mode = "aucroc"
func_evalmod_aucroc <- function(samp) {
  uaucs <- evalmod(scores = samp$scores, labels = samp$labels, mode="aucroc")
  as.data.frame(uaucs)
}

# Process time
system.time(res1 <- func_evalmod_rocprc(samp1))

#>    user  system elapsed 
#>   0.045   0.008   0.053 

system.time(res2 <- func_evalmod_aucroc(samp1))

#>    user  system elapsed 
#>   0.015   0.001   0.016 

# AUCs
res1

#>   modnames dsids curvetypes      aucs
#> 1       m1     1        ROC 0.5002722
#> 2       m1     1        PRC 0.4991961

res2

#>   modnames dsids      aucs     ustats
#> 1       m1     1 0.5002722 1250680392