Evaluate models and calculate performance evaluation measures

The evalmod function calculates ROC and Precision-Recall curves for specified prediction scores and binary labels. It also calculate several basic performance evaluation measures, such as accuracy, error rate, and precision, by specifying mode as "basic".

evalmod(mdat, mode = "rocprc", scores = NULL, labels = NULL,
  modnames = NULL, dsids = NULL, posclass = NULL, na_worst = TRUE,
  ties_method = "equiv", calc_avg = TRUE, cb_alpha = 0.05,
  raw_curves = FALSE, x_bins = 1000)

Arguments

mdat

An S3 object created by the mmdata function. It contains formatted scores and labels. The evalmod function ignores the following arguments when mdat is specified.

scores

labels

modnames

dsids

posclass

na_worst

ties_method

These arguments are internally passed to the mmdata function when mdat is unspecified. In that case, both scores and labels must be at least specified.

mode

A string that specifies the types of evaluation measures that the evalmod function calculates.

"rocprc"

ROC and Precision-Recall curves

"prcroc"

Same as above

"basic"

Normalized ranks vs. accuracy, error rate, specificity, sensitivity, precision, Matthews correlation coefficient, and F-score.

scores

A numeric dataset of predicted scores. It can be a vector, a matrix, an array, a data frame, or a list. The join_scores function can be useful to make scores with multiple datasets.

labels

A numeric, character, logical, or factor dataset of observed labels. It can be a vector, a matrix, an array, a data frame, or a list. The join_labels function can be useful to make labels with multiple datasets.

modnames

A character vector for the names of the models. The evalmod function automatically generates default names as "m1", "m2", "m3", and so on when it is NULL.

dsids

A numeric vector for test dataset IDs. The evalmod function automatically generates the default ID as 1 when it is NULL.

posclass

A scalar value to specify the label of positives in labels. It must be the same data type as labels. For example, posclass = -1 changes the positive label from 1 to -1 when labels contains 1 and -1. The positive label will be automatically detected when posclass is NULL.

na_worst

A Boolean value for controlling the treatment of NAs in scores.

TRUE

NAs are treated as the highest score

FALSE

NAs are treated as the lowest score

ties_method

A string for controlling ties in scores.

"equiv"

Ties are equivalently ranked

"first"

Ties are ranked in an increasing order as appeared

"random"

Ties are ranked in random order

calc_avg

A logical value to specify whether average curves should be calculated. It is effective only when dsids contains multiple dataset IDs. For instance, the function calculates the average for the model "m1" when modnames is c("m1", "m1", "m1") and dsids is c(1, 2, 3). The calculation points are defined by x_bins.

cb_alpha

A numeric value with range [0, 1] to specify the alpha value of the point-wise confidence bounds calculation. It is effective only when calc_avg is set to TRUE. For example, it should be 0.05 for the 95% confidence level. The calculation points are defined by x_bins.

raw_curves

A logical value to specify whether all raw curves should be discarded after the average curves are calculated. It is effective only when calc_avg is set to TRUE.

x_bins

An integer value to specify the number of minimum bins on the x-axis. It is then used to define supporting poins. For instance, the x-values of the supporting points will be c(0, 0.5, 1) and c(0, 0.25, 0.5, 0.75, 1) when x_bins = 2 and x_bins = 4, respectively. All corresponding y-values of the supporting points are calculated.

Value

The evalmod function returns an S3 object that contains performance evaluation measures. The number of models and the number of datasets can be controlled by modnames and dsids. For example, the number of models is "single" and the number of test datasets is "multiple" when modnames = c("m1", "m1", "m1") and dsids = c(1, 2, 3) are specified.

The evalmod function returns one of the following S3 objects when mode is "prcroc". The objects contain ROC and Precision-Recall curves.

S3 object 	 # of models

# of test datasets

sscurves
single 	 single

mscurves 	 multiple

single 	 smcurves

single 	 multiple

The evalmod function returns one of the following S3 objects when mode is "basic". They contain five different basic evaluation measures; error rate, accuracy, specificity, sensitivity, and precision.

S3 object 	 # of models

# of test datasets

sspoints
single 	 single

mspoints 	 multiple

single 	 smpoints

single 	 multiple

Different S3 objects have different default behaviors of S3 generics, such as plot, autoplot, and fortify.

See also

plot for plotting curves with the general R plot. autoplot and fortify for plotting curves with ggplot2. mmdata for formatting input data. join_scores and join_labels for formatting scores and labels with multiple datasets. create_sim_samples for generating random samples for simulations.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)
sscurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1         m1          1        ROC 0.7200000
#>    2         m1          1        PRC 0.7397716
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1         m1          1             10             10
#> 

## Generate an sspoints object that contains basic evaluation measures
sspoints <- evalmod(mode = "basic", scores = P10N10$scores,
                    labels = P10N10$labels)
sspoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>      Model ID  Meas.    Min. 1st Qu.  Median    Mean 3rd Qu.    Max.
#>    1    m1  1   rank  0.0000  0.2500  0.5000  0.5000  0.7500  1.0000
#>    2    m1  1  score  5.0000  5.7500 14.0000 11.7500 15.2500 20.0000
#>    3    m1  1  label -1.0000 -1.0000  0.0000  0.0000  1.0000  1.0000
#>    4    m1  1    err  0.3000  0.3500  0.4000  0.3952  0.4400  0.5000
#>    5    m1  1    acc  0.5000  0.5600  0.6000  0.6048  0.6500  0.7000
#>    6    m1  1     sp  0.0000  0.4000  0.6333  0.6048  0.9000  1.0000
#>    7    m1  1     sn  0.0000  0.4000  0.6333  0.6048  0.9000  1.0000
#>    8    m1  1   prec  0.5000  0.5750  0.6333  0.6892  0.7619  1.0000
#>    9    m1  1    mcc  0.1376  0.2238  0.2667  0.2756  0.3368  0.4364
#>   10    m1  1 fscore  0.0000  0.5333  0.6333  0.5580  0.6759  0.7200
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1         m1          1             10             10
#> 

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)
mscurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1     random          1        ROC 0.4991000
#>    2     random          1        PRC 0.5239905
#>    3    poor_er          1        ROC 0.7230000
#>    4    poor_er          1        PRC 0.6607375
#>    5    good_er          1        ROC 0.7423000
#>    6    good_er          1        PRC 0.7846787
#>    7      excel          1        ROC 0.9849000
#>    8      excel          1        PRC 0.9851427
#>    9       perf          1        ROC 1.0000000
#>   10       perf          1        PRC 1.0000000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#> 

## Generate an mspoints object that contains basic evaluation measures
mspoints <- evalmod(mdat, mode = "basic")
mspoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.      Min.  1st Qu.   Median      Mean 3rd Qu.   Max.
#>    1  random  1   rank  0.000000  0.25000  0.50000  0.500000 0.75000 1.0000
#>    2  random  1  score -3.031000 -0.71380 -0.05042 -0.069690 0.63450 2.5680
#>    3  random  1  label -1.000000 -1.00000  0.00000  0.000000 1.00000 1.0000
#>    4  random  1    err  0.465000  0.48500  0.50000  0.500400 0.51500 0.5400
#>    5  random  1    acc  0.460000  0.48500  0.50000  0.499600 0.51500 0.5350
#>    6  random  1     sp  0.000000  0.28000  0.49000  0.499600 0.74000 1.0000
#>    7  random  1     sn  0.000000  0.24000  0.49000  0.499600 0.78000 1.0000
#>    8  random  1   prec  0.436400  0.48310  0.50000  0.518100 0.51530 1.0000
#>    9  random  1    mcc -0.082010 -0.03039  0.00000  0.008462 0.05265 0.1255
#>   10  random  1 fscore  0.000000  0.31290  0.49250  0.449200 0.62150 0.6690
#>   11 poor_er  1   rank  0.000000  0.25000  0.50000  0.500000 0.75000 1.0000
#>   12 poor_er  1  score  0.008629  0.50380  0.75790  0.658500 0.87950 0.9981
#>   13 poor_er  1  label -1.000000 -1.00000  0.00000  0.000000 1.00000 1.0000
#>   14 poor_er  1    err  0.310000  0.34000  0.37500  0.389100 0.43500 0.5000
#>   15 poor_er  1    acc  0.500000  0.56500  0.62500  0.610900 0.66000 0.6900
#>   16 poor_er  1     sp  0.000000  0.42000  0.65000  0.610900 0.82000 1.0000
#>   17 poor_er  1     sn  0.000000  0.32000  0.65000  0.610900 0.92000 1.0000
#>   18 poor_er  1   prec  0.500000  0.60760  0.63640  0.638800 0.67390 1.0000
#>   19 poor_er  1    mcc  0.000000  0.17760  0.28580  0.267800 0.35290 0.4417
#>   20 poor_er  1 fscore  0.000000  0.43240  0.64580  0.559600 0.71620 0.7510
#>   21 good_er  1   rank  0.000000  0.25000  0.50000  0.500000 0.75000 1.0000
#>   22 good_er  1  score  0.002512  0.10360  0.28800  0.342100 0.51210 0.9877
#>   23 good_er  1  label -1.000000 -1.00000  0.00000  0.000000 1.00000 1.0000
#>   24 good_er  1    err  0.305000  0.35000  0.36000  0.379500 0.40500 0.5000
#>   25 good_er  1    acc  0.500000  0.59500  0.64000  0.620500 0.65000 0.6950
#>   26 good_er  1     sp  0.000000  0.41000  0.63000  0.620500 0.93000 1.0000
#>   27 good_er  1     sn  0.000000  0.43000  0.63000  0.620500 0.91000 1.0000
#>   28 good_er  1   prec  0.500000  0.60000  0.63390  0.712000 0.86270 1.0000
#>   29 good_er  1    mcc  0.070890  0.26310  0.29460  0.296200 0.34040 0.4670
#>   30 good_er  1 fscore  0.000000  0.56960  0.63270  0.582700 0.69260 0.7280
#>   31   excel  1   rank  0.000000  0.25000  0.50000  0.500000 0.75000 1.0000
#>   32   excel  1  score -2.812000 -0.04022  1.46600  1.514000 3.02100 5.4750
#>   33   excel  1  label -1.000000 -1.00000  0.00000  0.000000 1.00000 1.0000
#>   34   excel  1    err  0.070000  0.13000  0.25000  0.258800 0.37500 0.5000
#>   35   excel  1    acc  0.500000  0.62500  0.75000  0.741200 0.87000 0.9300
#>   36   excel  1     sp  0.000000  0.50000  0.92000  0.741200 1.00000 1.0000
#>   37   excel  1     sn  0.000000  0.50000  0.92000  0.741200 1.00000 1.0000
#>   38   excel  1   prec  0.500000  0.66670  0.91920  0.838400 1.00000 1.0000
#>   39   excel  1    mcc  0.070890  0.38230  0.57740  0.558300 0.76850 0.8686
#>   40   excel  1 fscore  0.000000  0.66670  0.76340  0.702800 0.87150 0.9346
#>   41    perf  1   rank  0.000000  0.25000  0.50000  0.500000 0.75000 1.0000
#>   42    perf  1  score  0.000000  0.00000  0.50000  0.500000 1.00000 1.0000
#>   43    perf  1  label -1.000000 -1.00000  0.00000  0.000000 1.00000 1.0000
#>   44    perf  1    err  0.000000  0.12500  0.25000  0.251200 0.37500 0.5000
#>   45    perf  1    acc  0.500000  0.62500  0.75000  0.748800 0.87500 1.0000
#>   46    perf  1     sp  0.000000  0.50000  1.00000  0.748800 1.00000 1.0000
#>   47    perf  1     sn  0.000000  0.50000  1.00000  0.748800 1.00000 1.0000
#>   48    perf  1   prec  0.500000  0.66670  1.00000  0.846100 1.00000 1.0000
#>   49    perf  1    mcc  0.070890  0.38230  0.57740  0.573500 0.77460 1.0000
#>   50    perf  1 fscore  0.000000  0.66670  0.76340  0.710400 0.87640 1.0000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#> 

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat)
smcurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1    good_er          1        ROC 0.8091000
#>    2    good_er          1        PRC 0.8415332
#>    3    good_er          2        ROC 0.8358000
#>    4    good_er          2        PRC 0.8715457
#>    5    good_er          3        ROC 0.7994000
#>    6    good_er          3        PRC 0.8387636
#>    7    good_er          4        ROC 0.8503000
#>    8    good_er          4        PRC 0.8864007
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1    good_er          1            100            100
#>    2    good_er          2            100            100
#>    3    good_er          3            100            100
#>    4    good_er          4            100            100
#> 

## Generate an smpoints object that contains basic evaluation measures
smpoints <- evalmod(mdat, mode = "basic")
smpoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.       Min.  1st Qu. Median   Mean 3rd Qu.   Max.
#>    1 good_er  1   rank  0.0000000  0.25000 0.5000 0.5000  0.7500 1.0000
#>    2 good_er  1  score  0.0001116  0.11750 0.2855 0.3585  0.5730 0.9991
#>    3 good_er  1  label -1.0000000 -1.00000 0.0000 0.0000  1.0000 1.0000
#>    4 good_er  1    err  0.2400000  0.29000 0.3300 0.3462  0.3950 0.5000
#>    5 good_er  1    acc  0.5000000  0.60500 0.6700 0.6538  0.7100 0.7600
#>    6 good_er  1     sp  0.0000000  0.40000 0.7000 0.6538  0.9700 1.0000
#>    7 good_er  1     sn  0.0000000  0.47000 0.7000 0.6538  0.9000 1.0000
#>    8 good_er  1   prec  0.5000000  0.60000 0.7059 0.7505  0.9333 1.0000
#>    9 good_er  1    mcc  0.0708900  0.30000 0.3802 0.3696  0.4375 0.5385
#>   10 good_er  1 fscore  0.0000000  0.61840 0.7019 0.6175  0.7156 0.7325
#>   11 good_er  2   rank  0.0000000  0.25000 0.5000 0.5000  0.7500 1.0000
#>   12 good_er  2  score  0.0017660  0.08911 0.2376 0.3329  0.5459 0.9765
#>   13 good_er  2  label -1.0000000 -1.00000 0.0000 0.0000  1.0000 1.0000
#>   14 good_er  2    err  0.2100000  0.26000 0.3200 0.3329  0.4000 0.5000
#>   15 good_er  2    acc  0.5000000  0.60000 0.6800 0.6671  0.7400 0.7900
#>   16 good_er  2     sp  0.0000000  0.40000 0.7400 0.6671  0.9700 1.0000
#>   17 good_er  2     sn  0.0000000  0.47000 0.7400 0.6671  0.9000 1.0000
#>   18 good_er  2   prec  0.5000000  0.59730 0.7426 0.7684  0.9388 1.0000
#>   19 good_er  2    mcc  0.0000000  0.30720 0.4312 0.3920  0.5100 0.5842
#>   20 good_er  2 fscore  0.0000000  0.62670 0.7104 0.6325  0.7426 0.7766
#>   21 good_er  3   rank  0.0000000  0.25000 0.5000 0.5000  0.7500 1.0000
#>   22 good_er  3  score  0.0012830  0.11280 0.2621 0.3450  0.5068 0.9894
#>   23 good_er  3  label -1.0000000 -1.00000 0.0000 0.0000  1.0000 1.0000
#>   24 good_er  3    err  0.2650000  0.29500 0.3250 0.3510  0.4100 0.5000
#>   25 good_er  3    acc  0.5000000  0.59000 0.6750 0.6490  0.7050 0.7350
#>   26 good_er  3     sp  0.0000000  0.41000 0.7000 0.6490  0.9600 1.0000
#>   27 good_er  3     sn  0.0000000  0.46000 0.7000 0.6490  0.9100 1.0000
#>   28 good_er  3   prec  0.5000000  0.60670 0.7000 0.7477  0.9184 1.0000
#>   29 good_er  3    mcc  0.0708900  0.27240 0.3767 0.3564  0.4596 0.5078
#>   30 good_er  3 fscore  0.0000000  0.61330 0.6879 0.6135  0.7091 0.7382
#>   31 good_er  4   rank  0.0000000  0.25000 0.5000 0.5000  0.7500 1.0000
#>   32 good_er  4  score  0.0044120  0.12670 0.2868 0.3545  0.5392 0.9925
#>   33 good_er  4  label -1.0000000 -1.00000 0.0000 0.0000  1.0000 1.0000
#>   34 good_er  4    err  0.2050000  0.23500 0.3300 0.3257  0.4000 0.5000
#>   35 good_er  4    acc  0.5000000  0.60000 0.6700 0.6743  0.7650 0.7950
#>   36 good_er  4     sp  0.0000000  0.41000 0.7900 0.6743  1.0000 1.0000
#>   37 good_er  4     sn  0.0000000  0.50000 0.7900 0.6743  0.9100 1.0000
#>   38 good_er  4   prec  0.5000000  0.60140 0.7857 0.7770  1.0000 1.0000
#>   39 good_er  4    mcc  0.0708900  0.30920 0.4118 0.4128  0.5553 0.6092
#>   40 good_er  4 fscore  0.0000000  0.66670 0.7131 0.6399  0.7381 0.7900
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1    good_er          1            100            100
#>    2    good_er          2            100            100
#>    3    good_er          3            100            100
#>    4    good_er          4            100            100
#> 

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat)
mmcurves

#> 
#>     === AUCs ===
#> 
#>      Model name Dataset ID Curve type       AUC
#>    1     random          1        ROC 0.4537000
#>    2     random          1        PRC 0.4575203
#>    3    poor_er          1        ROC 0.7563000
#>    4    poor_er          1        PRC 0.6761757
#>    5    good_er          1        ROC 0.8135000
#>    6    good_er          1        PRC 0.8358523
#>    7      excel          1        ROC 0.9893000
#>    8      excel          1        PRC 0.9898202
#>    9       perf          1        ROC 1.0000000
#>   10       perf          1        PRC 1.0000000
#>   11     random          2        ROC 0.5561000
#>   12     random          2        PRC 0.5481585
#>   13    poor_er          2        ROC 0.8474000
#>   14    poor_er          2        PRC 0.8310307
#>   15    good_er          2        ROC 0.8088000
#>   16    good_er          2        PRC 0.8377002
#>   17      excel          2        ROC 0.9818000
#>   18      excel          2        PRC 0.9801307
#>   19       perf          2        ROC 1.0000000
#>   20       perf          2        PRC 1.0000000
#>   21     random          3        ROC 0.5626000
#>   22     random          3        PRC 0.5281464
#>   23    poor_er          3        ROC 0.7887000
#>   24    poor_er          3        PRC 0.7416898
#>   25    good_er          3        ROC 0.8048000
#>   26    good_er          3        PRC 0.8509163
#>   27      excel          3        ROC 0.9583000
#>   28      excel          3        PRC 0.9529454
#>   29       perf          3        ROC 1.0000000
#>   30       perf          3        PRC 1.0000000
#>   31     random          4        ROC 0.4650000
#>   32     random          4        PRC 0.4903973
#>   33    poor_er          4        ROC 0.8157000
#>   34    poor_er          4        PRC 0.7646248
#>   35    good_er          4        ROC 0.7897000
#>   36    good_er          4        PRC 0.8278571
#>   37      excel          4        ROC 0.9843000
#>   38      excel          4        PRC 0.9865355
#>   39       perf          4        ROC 1.0000000
#>   40       perf          4        PRC 1.0000000
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#>    6     random          2            100            100
#>    7    poor_er          2            100            100
#>    8    good_er          2            100            100
#>    9      excel          2            100            100
#>   10       perf          2            100            100
#>   11     random          3            100            100
#>   12    poor_er          3            100            100
#>   13    good_er          3            100            100
#>   14      excel          3            100            100
#>   15       perf          3            100            100
#>   16     random          4            100            100
#>   17    poor_er          4            100            100
#>   18    good_er          4            100            100
#>   19      excel          4            100            100
#>   20       perf          4            100            100
#> 

## Generate an mmpoints object that contains basic evaluation measures
mmpoints <- evalmod(mdat, mode = "basic")
mmpoints

#> 
#>     === Basic performance evaluation measures ===
#> 
#>      ## Performance measures (Meas.)
#>       rank:   normalized rank
#>       score:  score
#>       label:  label
#>       err:    error rate
#>       acc:    accuracy
#>       sp:     specificity
#>       sn:     sensitivity
#>       prec:   precision
#>       mcc:    Matthews correlation coefficient
#>       fscore: F-score
#> 
#> 
#>        Model ID  Meas.       Min.  1st Qu.   Median      Mean  3rd Qu.   Max.
#>    1  random  1   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>    2  random  1  score -3.366e+00 -0.63840 -0.06593 -0.002629  0.71970 2.7710
#>    3  random  1  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>    4  random  1    err  4.750e-01  0.51000  0.53000  0.523000  0.53500 0.5550
#>    5  random  1    acc  4.450e-01  0.46500  0.47000  0.477000  0.49000 0.5250
#>    6  random  1     sp  0.000e+00  0.23000  0.45000  0.477000  0.71000 1.0000
#>    7  random  1     sn  0.000e+00  0.21000  0.45000  0.477000  0.73000 1.0000
#>    8  random  1   prec  0.000e+00  0.44440  0.46220  0.449600  0.48650 0.5131
#>    9  random  1    mcc -1.255e-01 -0.08203 -0.06392 -0.051900 -0.03025 0.1206
#>   10  random  1 fscore  0.000e+00  0.28380  0.45000  0.423700  0.58400 0.6735
#>   11 poor_er  1   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   12 poor_er  1  score  1.534e-03  0.55670  0.73070  0.671000  0.87700 0.9968
#>   13 poor_er  1  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   14 poor_er  1    err  2.500e-01  0.31500  0.36500  0.372500  0.43000 0.5050
#>   15 poor_er  1    acc  4.950e-01  0.57000  0.63500  0.627500  0.68500 0.7500
#>   16 poor_er  1     sp  0.000e+00  0.45000  0.67000  0.627500  0.85000 1.0000
#>   17 poor_er  1     sn  0.000e+00  0.35000  0.67000  0.627500  0.95000 1.0000
#>   18 poor_er  1   prec  3.333e-01  0.61540  0.67310  0.649900  0.68750 1.0000
#>   19 poor_er  1    mcc -4.113e-02  0.21910  0.30220  0.302300  0.42160 0.5359
#>   20 poor_er  1 fscore  0.000e+00  0.46670  0.66670  0.575200  0.73640 0.7881
#>   21 good_er  1   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   22 good_er  1  score  2.891e-03  0.11090  0.28440  0.353600  0.56970 0.9969
#>   23 good_er  1  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   24 good_er  1    err  2.150e-01  0.26000  0.33000  0.344000  0.42500 0.5000
#>   25 good_er  1    acc  5.000e-01  0.57500  0.67000  0.656000  0.74000 0.7850
#>   26 good_er  1     sp  0.000e+00  0.40000  0.76000  0.656000  0.95000 1.0000
#>   27 good_er  1     sn  0.000e+00  0.45000  0.76000  0.656000  0.90000 1.0000
#>   28 good_er  1   prec  5.000e-01  0.60000  0.75760  0.748400  0.90000 1.0000
#>   29 good_er  1    mcc  0.000e+00  0.24570  0.38010  0.363900  0.50170 0.5784
#>   30 good_er  1 fscore  0.000e+00  0.60000  0.68940  0.619800  0.72880 0.7692
#>   31   excel  1   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   32   excel  1  score -2.287e+00  0.05779  1.65800  1.522000  2.99000 5.6900
#>   33   excel  1  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   34   excel  1    err  4.000e-02  0.13000  0.25000  0.256600  0.37500 0.5000
#>   35   excel  1    acc  5.000e-01  0.62500  0.75000  0.743400  0.87000 0.9600
#>   36   excel  1     sp  0.000e+00  0.50000  0.95000  0.743400  1.00000 1.0000
#>   37   excel  1     sn  0.000e+00  0.50000  0.95000  0.743400  1.00000 1.0000
#>   38   excel  1   prec  5.000e-01  0.66670  0.94950  0.840700  1.00000 1.0000
#>   39   excel  1    mcc  7.089e-02  0.38230  0.57740  0.562700  0.76230 0.9207
#>   40   excel  1 fscore  0.000e+00  0.66670  0.76340  0.705100  0.86960 0.9608
#>   41    perf  1   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   42    perf  1  score  0.000e+00  0.00000  0.50000  0.500000  1.00000 1.0000
#>   43    perf  1  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   44    perf  1    err  0.000e+00  0.12500  0.25000  0.251200  0.37500 0.5000
#>   45    perf  1    acc  5.000e-01  0.62500  0.75000  0.748800  0.87500 1.0000
#>   46    perf  1     sp  0.000e+00  0.50000  1.00000  0.748800  1.00000 1.0000
#>   47    perf  1     sn  0.000e+00  0.50000  1.00000  0.748800  1.00000 1.0000
#>   48    perf  1   prec  5.000e-01  0.66670  1.00000  0.846100  1.00000 1.0000
#>   49    perf  1    mcc  7.089e-02  0.38230  0.57740  0.573500  0.77460 1.0000
#>   50    perf  1 fscore  0.000e+00  0.66670  0.76340  0.710400  0.87640 1.0000
#>   51  random  2   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   52  random  2  score -2.279e+00 -0.58310 -0.07525  0.042210  0.66200 2.7000
#>   53  random  2  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   54  random  2    err  4.250e-01  0.46000  0.47500  0.472100  0.48500 0.5100
#>   55  random  2    acc  4.900e-01  0.51500  0.52500  0.527900  0.54000 0.5750
#>   56  random  2     sp  0.000e+00  0.26000  0.54000  0.527900  0.80000 1.0000
#>   57  random  2     sn  0.000e+00  0.30000  0.54000  0.527900  0.76000 1.0000
#>   58  random  2   prec  0.000e+00  0.50980  0.53540  0.540200  0.58330 0.6667
#>   59  random  2    mcc -7.143e-02  0.03594  0.07000  0.066130  0.09664 0.1601
#>   60  random  2 fscore  0.000e+00  0.39470  0.54000  0.481800  0.61040 0.6689
#>   61 poor_er  2   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   62 poor_er  2  score  3.613e-03  0.39560  0.68710  0.623600  0.84300 0.9996
#>   63 poor_er  2  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   64 poor_er  2    err  2.100e-01  0.25000  0.31000  0.327200  0.39500 0.5000
#>   65 poor_er  2    acc  5.000e-01  0.60500  0.69000  0.672800  0.75000 0.7900
#>   66 poor_er  2     sp  0.000e+00  0.45000  0.79000  0.672800  0.92000 1.0000
#>   67 poor_er  2     sn  0.000e+00  0.42000  0.79000  0.672800  0.95000 1.0000
#>   68 poor_er  2   prec  5.000e-01  0.63330  0.78490  0.750300  0.84780 1.0000
#>   69 poor_er  2    mcc  7.089e-02  0.32570  0.44170  0.409400  0.51560 0.5801
#>   70 poor_er  2 fscore  0.000e+00  0.56000  0.71910  0.630000  0.76920 0.7900
#>   71 good_er  2   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   72 good_er  2  score  3.891e-03  0.12570  0.28030  0.351100  0.50730 0.9972
#>   73 good_er  2  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   74 good_er  2    err  2.400e-01  0.27500  0.33500  0.346400  0.40500 0.5000
#>   75 good_er  2    acc  5.000e-01  0.59500  0.66500  0.653600  0.72500 0.7600
#>   76 good_er  2     sp  0.000e+00  0.38000  0.74000  0.653600  0.95000 1.0000
#>   77 good_er  2     sn  0.000e+00  0.45000  0.74000  0.653600  0.88000 1.0000
#>   78 good_er  2   prec  5.000e-01  0.58280  0.74000  0.748800  0.89580 1.0000
#>   79 good_er  2    mcc  7.089e-02  0.27590  0.39920  0.364600  0.47170 0.5331
#>   80 good_er  2 fscore  0.000e+00  0.60000  0.69750  0.617200  0.72340 0.7619
#>   81   excel  2   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   82   excel  2  score -2.807e+00  0.27070  1.63200  1.523000  2.95800 4.8110
#>   83   excel  2  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   84   excel  2    err  6.000e-02  0.13000  0.25500  0.260300  0.37500 0.5000
#>   85   excel  2    acc  5.000e-01  0.62500  0.74500  0.739700  0.87000 0.9400
#>   86   excel  2     sp  0.000e+00  0.50000  0.94000  0.739700  0.99000 1.0000
#>   87   excel  2     sn  0.000e+00  0.49000  0.94000  0.739700  1.00000 1.0000
#>   88   excel  2   prec  5.000e-01  0.66670  0.93940  0.835800  0.98630 1.0000
#>   89   excel  2    mcc  7.089e-02  0.38230  0.56970  0.554800  0.76430 0.8807
#>   90   excel  2 fscore  0.000e+00  0.65330  0.76050  0.700900  0.87150 0.9406
#>   91    perf  2   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>   92    perf  2  score  0.000e+00  0.00000  0.50000  0.500000  1.00000 1.0000
#>   93    perf  2  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>   94    perf  2    err  0.000e+00  0.12500  0.25000  0.251200  0.37500 0.5000
#>   95    perf  2    acc  5.000e-01  0.62500  0.75000  0.748800  0.87500 1.0000
#>   96    perf  2     sp  0.000e+00  0.50000  1.00000  0.748800  1.00000 1.0000
#>   97    perf  2     sn  0.000e+00  0.50000  1.00000  0.748800  1.00000 1.0000
#>   98    perf  2   prec  5.000e-01  0.66670  1.00000  0.846100  1.00000 1.0000
#>   99    perf  2    mcc  7.089e-02  0.38230  0.57740  0.573500  0.77460 1.0000
#>  100    perf  2 fscore  0.000e+00  0.66670  0.76340  0.710400  0.87640 1.0000
#>  101  random  3   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>  102  random  3  score -2.457e+00 -0.79070 -0.06154 -0.081890  0.60420 1.7680
#>  103  random  3  label -1.000e+00 -1.00000  0.00000  0.000000  1.00000 1.0000
#>  104  random  3    err  4.150e-01  0.44500  0.47000  0.468900  0.49500 0.5100
#>  105  random  3    acc  4.900e-01  0.50500  0.53000  0.531100  0.55500 0.5850
#>  106  random  3     sp  0.000e+00  0.32000  0.55000  0.531100  0.76000 1.0000
#>  107  random  3     sn  0.000e+00  0.26000  0.55000  0.531100  0.82000 1.0000
#>  108  random  3   prec  0.000e+00  0.50850  0.53470  0.522000  0.54640 0.6667
#>  109  random  3    mcc -1.005e-01  0.01488  0.06124  0.067930  0.12810 0.1796
#>  110  random  3 fscore  0.000e+00  0.34670  0.54730  0.479100  0.64750 0.6667
#>  111 poor_er  3   rank  0.000e+00  0.25000  0.50000  0.500000  0.75000 1.0000
#>  [ reached getOption("max.print") -- omitted 89 rows ]
#> 
#> 
#>     === Input data ===
#> 
#>      Model name Dataset ID # of positives # of negatives
#>    1     random          1            100            100
#>    2    poor_er          1            100            100
#>    3    good_er          1            100            100
#>    4      excel          1            100            100
#>    5       perf          1            100            100
#>    6     random          2            100            100
#>    7    poor_er          2            100            100
#>    8    good_er          2            100            100
#>    9      excel          2            100            100
#>   10       perf          2            100            100
#>   11     random          3            100            100
#>   12    poor_er          3            100            100
#>   13    good_er          3            100            100
#>   14      excel          3            100            100
#>   15       perf          3            100            100
#>   16     random          4            100            100
#>   17    poor_er          4            100            100
#>   18    good_er          4            100            100
#>   19      excel          4            100            100
#>   20       perf          4            100            100
#>