Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves

An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object 	 # of models

# of test datasets 	 sscurves

single 	 single

mscurves 	 multiple

single 	 smcurves

single 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. part for calculation of pAUCs. auc for retieving a dataset of AUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1228000 0.3274667
#> 2    random     1        PRC 0.1947165 0.5192441
#> 3   poor_er     1        ROC 0.3065000 0.8173333
#> 4   poor_er     1        PRC 0.3750000 1.0000000
#> 5   good_er     1        ROC 0.3210000 0.8560000
#> 6   good_er     1        PRC 0.3679145 0.9811054
#> 7     excel     1        ROC 0.3716000 0.9909333
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(10, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1   good_er     1        ROC 0.4142000 0.8284000
#> 2   good_er     1        PRC 0.4514434 0.9028868
#> 3   good_er     2        ROC 0.4139000 0.8278000
#> 4   good_er     2        PRC 0.4295945 0.8591890
#> 5   good_er     3        ROC 0.4314000 0.8628000
#> 6   good_er     3        PRC 0.4555022 0.9110043
#> 7   good_er     4        ROC 0.4147000 0.8294000
#> 8   good_er     4        PRC 0.4468420 0.8936839
#> 9   good_er     5        ROC 0.4026000 0.8052000
#> 10  good_er     5        PRC 0.4066804 0.8133608
#> 11  good_er     6        ROC 0.4313000 0.8626000
#> 12  good_er     6        PRC 0.4463786 0.8927572
#> 13  good_er     7        ROC 0.4237000 0.8474000
#> 14  good_er     7        PRC 0.4364526 0.8729052
#> 15  good_er     8        ROC 0.3912000 0.7824000
#> 16  good_er     8        PRC 0.4175689 0.8351377
#> 17  good_er     9        ROC 0.4251000 0.8502000
#> 18  good_er     9        PRC 0.4413570 0.8827141
#> 19  good_er    10        ROC 0.3869000 0.7738000
#> 20  good_er    10        PRC 0.4249428 0.8498855

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(10, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)

#>     modnames dsids curvetypes      paucs    spaucs
#> 1     random     1        ROC 0.01540000 0.0616000
#> 2     random     1        PRC 0.08752726 0.3501091
#> 3    poor_er     1        ROC 0.08540000 0.3416000
#> 4    poor_er     1        PRC 0.17666449 0.7066579
#> 5    good_er     1        ROC 0.14430000 0.5772000
#> 6    good_er     1        PRC 0.25000000 1.0000000
#> 7      excel     1        ROC 0.23080000 0.9232000
#> 8      excel     1        PRC 0.25000000 1.0000000
#> 9       perf     1        ROC 0.25000000 1.0000000
#> 10      perf     1        PRC 0.25000000 1.0000000
#> 11    random     2        ROC 0.02520000 0.1008000
#> 12    random     2        PRC 0.11451106 0.4580443
#> 13   poor_er     2        ROC 0.07510000 0.3004000
#> 14   poor_er     2        PRC 0.17323749 0.6929499
#> 15   good_er     2        ROC 0.13050000 0.5220000
#> 16   good_er     2        PRC 0.25000000 1.0000000
#> 17     excel     2        ROC 0.23300000 0.9320000
#> 18     excel     2        PRC 0.25000000 1.0000000
#> 19      perf     2        ROC 0.25000000 1.0000000
#> 20      perf     2        PRC 0.25000000 1.0000000
#> 21    random     3        ROC 0.03150000 0.1260000
#> 22    random     3        PRC 0.12316272 0.4926509
#> 23   poor_er     3        ROC 0.07220000 0.2888000
#> 24   poor_er     3        PRC 0.17810947 0.7124379
#> 25   good_er     3        ROC 0.14820000 0.5928000
#> 26   good_er     3        PRC 0.25000000 1.0000000
#> 27     excel     3        ROC 0.22600000 0.9040000
#> 28     excel     3        PRC 0.25000000 1.0000000
#> 29      perf     3        ROC 0.25000000 1.0000000
#> 30      perf     3        PRC 0.25000000 1.0000000
#> 31    random     4        ROC 0.04830000 0.1932000
#> 32    random     4        PRC 0.17432026 0.6972811
#> 33   poor_er     4        ROC 0.09680000 0.3872000
#> 34   poor_er     4        PRC 0.22237158 0.8894863
#> 35   good_er     4        ROC 0.13620000 0.5448000
#> 36   good_er     4        PRC 0.25000000 1.0000000
#> 37     excel     4        ROC 0.23820000 0.9528000
#> 38     excel     4        PRC 0.25000000 1.0000000
#> 39      perf     4        ROC 0.25000000 1.0000000
#> 40      perf     4        PRC 0.25000000 1.0000000
#> 41    random     5        ROC 0.01420000 0.0568000
#> 42    random     5        PRC 0.09084579 0.3633831
#> 43   poor_er     5        ROC 0.09560000 0.3824000
#> 44   poor_er     5        PRC 0.16691615 0.6676646
#> 45   good_er     5        ROC 0.12970000 0.5188000
#> 46   good_er     5        PRC 0.25000000 1.0000000
#> 47     excel     5        ROC 0.23430000 0.9372000
#> 48     excel     5        PRC 0.25000000 1.0000000
#> 49      perf     5        ROC 0.25000000 1.0000000
#> 50      perf     5        PRC 0.25000000 1.0000000
#> 51    random     6        ROC 0.03520000 0.1408000
#> 52    random     6        PRC 0.14035993 0.5614397
#> 53   poor_er     6        ROC 0.10130000 0.4052000
#> 54   poor_er     6        PRC 0.19541334 0.7816534
#> 55   good_er     6        ROC 0.14220000 0.5688000
#> 56   good_er     6        PRC 0.25000000 1.0000000
#> 57     excel     6        ROC 0.22840000 0.9136000
#> 58     excel     6        PRC 0.25000000 1.0000000
#> 59      perf     6        ROC 0.25000000 1.0000000
#> 60      perf     6        PRC 0.25000000 1.0000000
#> 61    random     7        ROC 0.04650000 0.1860000
#> 62    random     7        PRC 0.16169302 0.6467721
#> 63   poor_er     7        ROC 0.10290000 0.4116000
#> 64   poor_er     7        PRC 0.22964772 0.9185909
#> 65   good_er     7        ROC 0.15000000 0.6000000
#> 66   good_er     7        PRC 0.24044481 0.9617793
#> 67     excel     7        ROC 0.23510000 0.9404000
#> 68     excel     7        PRC 0.25000000 1.0000000
#> 69      perf     7        ROC 0.25000000 1.0000000
#> 70      perf     7        PRC 0.25000000 1.0000000
#> 71    random     8        ROC 0.05990000 0.2396000
#> 72    random     8        PRC 0.20643045 0.8257218
#> 73   poor_er     8        ROC 0.10950000 0.4380000
#> 74   poor_er     8        PRC 0.19253464 0.7701386
#> 75   good_er     8        ROC 0.11010000 0.4404000
#> 76   good_er     8        PRC 0.24737635 0.9895054
#> 77     excel     8        ROC 0.23120000 0.9248000
#> 78     excel     8        PRC 0.25000000 1.0000000
#> 79      perf     8        ROC 0.25000000 1.0000000
#> 80      perf     8        PRC 0.25000000 1.0000000
#> 81    random     9        ROC 0.03010000 0.1204000
#> 82    random     9        PRC 0.12509479 0.5003792
#> 83   poor_er     9        ROC 0.07860000 0.3144000
#> 84   poor_er     9        PRC 0.19975979 0.7990392
#> 85   good_er     9        ROC 0.13620000 0.5448000
#> 86   good_er     9        PRC 0.25000000 1.0000000
#> 87     excel     9        ROC 0.22100000 0.8840000
#> 88     excel     9        PRC 0.25000000 1.0000000
#> 89      perf     9        ROC 0.25000000 1.0000000
#> 90      perf     9        PRC 0.25000000 1.0000000
#> 91    random    10        ROC 0.02210000 0.0884000
#> 92    random    10        PRC 0.09916342 0.3966537
#> 93   poor_er    10        ROC 0.10230000 0.4092000
#> 94   poor_er    10        PRC 0.21467488 0.8586995
#> 95   good_er    10        ROC 0.15140000 0.6056000
#> 96   good_er    10        PRC 0.25000000 1.0000000
#> 97     excel    10        ROC 0.23300000 0.9320000
#> 98     excel    10        PRC 0.25000000 1.0000000
#> 99      perf    10        ROC 0.25000000 1.0000000
#> 100     perf    10        PRC 0.25000000 1.0000000