Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves
An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object	# of models	# of test datasets

sscurves	single	single

mscurves	multiple	single

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1249000 0.3330667
#> 2    random     1        PRC 0.1936290 0.5163439
#> 3   poor_er     1        ROC 0.3067000 0.8178667
#> 4   poor_er     1        PRC 0.3565796 0.9508789
#> 5   good_er     1        ROC 0.3111000 0.8296000
#> 6   good_er     1        PRC 0.3664967 0.9773245
#> 7     excel     1        ROC 0.3750000 1.0000000
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.4346000 0.8692000
#> 2  good_er     1        PRC 0.4529028 0.9058057
#> 3  good_er     2        ROC 0.4283000 0.8566000
#> 4  good_er     2        PRC 0.4627090 0.9254181
#> 5  good_er     3        ROC 0.4057000 0.8114000
#> 6  good_er     3        PRC 0.4147117 0.8294234
#> 7  good_er     4        ROC 0.4360000 0.8720000
#> 8  good_er     4        PRC 0.4361260 0.8722520

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.0319000 0.1276000
#> 2    random     1        PRC 0.1234931 0.4939725
#> 3   poor_er     1        ROC 0.0942000 0.3768000
#> 4   poor_er     1        PRC 0.2066198 0.8264793
#> 5   good_er     1        ROC 0.1393000 0.5572000
#> 6   good_er     1        PRC 0.2473763 0.9895054
#> 7     excel     1        ROC 0.2311000 0.9244000
#> 8     excel     1        PRC 0.2500000 1.0000000
#> 9      perf     1        ROC 0.2500000 1.0000000
#> 10     perf     1        PRC 0.2500000 1.0000000
#> 11   random     2        ROC 0.0362000 0.1448000
#> 12   random     2        PRC 0.1251075 0.5004301
#> 13  poor_er     2        ROC 0.1290000 0.5160000
#> 14  poor_er     2        PRC 0.2427651 0.9710605
#> 15  good_er     2        ROC 0.1360000 0.5440000
#> 16  good_er     2        PRC 0.2500000 1.0000000
#> 17    excel     2        ROC 0.2226000 0.8904000
#> 18    excel     2        PRC 0.2500000 1.0000000
#> 19     perf     2        ROC 0.2500000 1.0000000
#> 20     perf     2        PRC 0.2500000 1.0000000
#> 21   random     3        ROC 0.0348000 0.1392000
#> 22   random     3        PRC 0.1266683 0.5066733
#> 23  poor_er     3        ROC 0.0666000 0.2664000
#> 24  poor_er     3        PRC 0.1656674 0.6626698
#> 25  good_er     3        ROC 0.1330000 0.5320000
#> 26  good_er     3        PRC 0.2500000 1.0000000
#> 27    excel     3        ROC 0.2249000 0.8996000
#> 28    excel     3        PRC 0.2500000 1.0000000
#> 29     perf     3        ROC 0.2500000 1.0000000
#> 30     perf     3        PRC 0.2500000 1.0000000
#> 31   random     4        ROC 0.0246000 0.0984000
#> 32   random     4        PRC 0.1028529 0.4114116
#> 33  poor_er     4        ROC 0.0755000 0.3020000
#> 34  poor_er     4        PRC 0.1874306 0.7497224
#> 35  good_er     4        ROC 0.1298000 0.5192000
#> 36  good_er     4        PRC 0.2500000 1.0000000
#> 37    excel     4        ROC 0.2276000 0.9104000
#> 38    excel     4        PRC 0.2500000 1.0000000
#> 39     perf     4        ROC 0.2500000 1.0000000
#> 40     perf     4        PRC 0.2500000 1.0000000