Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves

An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object	# of models	# of test datasets

sscurves	single	single

mscurves	multiple	single

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)
#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1426000 0.3802667
#> 2    random     1        PRC 0.2324886 0.6199696
#> 3   poor_er     1        ROC 0.3034000 0.8090667
#> 4   poor_er     1        PRC 0.3565188 0.9507168
#> 5   good_er     1        ROC 0.3262000 0.8698667
#> 6   good_er     1        PRC 0.3736260 0.9963360
#> 7     excel     1        ROC 0.3711000 0.9896000
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.4180000 0.8360000
#> 2  good_er     1        PRC 0.4453282 0.8906565
#> 3  good_er     2        ROC 0.3947000 0.7894000
#> 4  good_er     2        PRC 0.4063834 0.8127668
#> 5  good_er     3        ROC 0.4186000 0.8372000
#> 6  good_er     3        PRC 0.4554943 0.9109886
#> 7  good_er     4        ROC 0.4119000 0.8238000
#> 8  good_er     4        PRC 0.4269404 0.8538808

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)
#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.0347000 0.1388000
#> 2    random     1        PRC 0.1377804 0.5511217
#> 3   poor_er     1        ROC 0.0822000 0.3288000
#> 4   poor_er     1        PRC 0.1986673 0.7946694
#> 5   good_er     1        ROC 0.1402000 0.5608000
#> 6   good_er     1        PRC 0.2500000 1.0000000
#> 7     excel     1        ROC 0.2356000 0.9424000
#> 8     excel     1        PRC 0.2500000 1.0000000
#> 9      perf     1        ROC 0.2500000 1.0000000
#> 10     perf     1        PRC 0.2500000 1.0000000
#> 11   random     2        ROC 0.0209000 0.0836000
#> 12   random     2        PRC 0.1169066 0.4676265
#> 13  poor_er     2        ROC 0.0552000 0.2208000
#> 14  poor_er     2        PRC 0.1766030 0.7064119
#> 15  good_er     2        ROC 0.1101000 0.4404000
#> 16  good_er     2        PRC 0.2450124 0.9800494
#> 17    excel     2        ROC 0.2343000 0.9372000
#> 18    excel     2        PRC 0.2500000 1.0000000
#> 19     perf     2        ROC 0.2500000 1.0000000
#> 20     perf     2        PRC 0.2500000 1.0000000
#> 21   random     3        ROC 0.0305000 0.1220000
#> 22   random     3        PRC 0.1350893 0.5403572
#> 23  poor_er     3        ROC 0.1359000 0.5436000
#> 24  poor_er     3        PRC 0.2365843 0.9463373
#> 25  good_er     3        ROC 0.1307000 0.5228000
#> 26  good_er     3        PRC 0.2478643 0.9914570
#> 27    excel     3        ROC 0.2318000 0.9272000
#> 28    excel     3        PRC 0.2500000 1.0000000
#> 29     perf     3        ROC 0.2500000 1.0000000
#> 30     perf     3        PRC 0.2500000 1.0000000
#> 31   random     4        ROC 0.0406000 0.1624000
#> 32   random     4        PRC 0.1458424 0.5833694
#> 33  poor_er     4        ROC 0.0831000 0.3324000
#> 34  poor_er     4        PRC 0.2086771 0.8347086
#> 35  good_er     4        ROC 0.1497000 0.5988000
#> 36  good_er     4        PRC 0.2500000 1.0000000
#> 37    excel     4        ROC 0.2260000 0.9040000
#> 38    excel     4        PRC 0.2500000 1.0000000
#> 39     perf     4        ROC 0.2500000 1.0000000
#> 40     perf     4        PRC 0.2500000 1.0000000