Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves

An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object 	 # of models 	 # of test datasets

sscurves 	 single 	 single

mscurves 	 multiple 	 single

smcurves 	 single 	 multiple

mmcurves 	 multiple 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. part for calculation of pAUCs. auc for retieving a dataset of AUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1528000 0.4074667
#> 2    random     1        PRC 0.2248972 0.5997258
#> 3   poor_er     1        ROC 0.2490000 0.6640000
#> 4   poor_er     1        PRC 0.3215164 0.8573772
#> 5   good_er     1        ROC 0.3312000 0.8832000
#> 6   good_er     1        PRC 0.3742979 0.9981276
#> 7     excel     1        ROC 0.3722000 0.9925333
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.3858000 0.7716000
#> 2  good_er     1        PRC 0.3987139 0.7974278
#> 3  good_er     2        ROC 0.4270000 0.8540000
#> 4  good_er     2        PRC 0.4573763 0.9147525
#> 5  good_er     3        ROC 0.4186000 0.8372000
#> 6  good_er     3        PRC 0.4718847 0.9437694
#> 7  good_er     4        ROC 0.4307000 0.8614000
#> 8  good_er     4        PRC 0.4672641 0.9345281

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.0332000 0.1328000
#> 2    random     1        PRC 0.1343677 0.5374709
#> 3   poor_er     1        ROC 0.1179000 0.4716000
#> 4   poor_er     1        PRC 0.2178800 0.8715200
#> 5   good_er     1        ROC 0.1285000 0.5140000
#> 6   good_er     1        PRC 0.2496078 0.9984312
#> 7     excel     1        ROC 0.2271000 0.9084000
#> 8     excel     1        PRC 0.2500000 1.0000000
#> 9      perf     1        ROC 0.2500000 1.0000000
#> 10     perf     1        PRC 0.2500000 1.0000000
#> 11   random     2        ROC 0.0340000 0.1360000
#> 12   random     2        PRC 0.1190802 0.4763208
#> 13  poor_er     2        ROC 0.1228000 0.4912000
#> 14  poor_er     2        PRC 0.2380826 0.9523303
#> 15  good_er     2        ROC 0.1487000 0.5948000
#> 16  good_er     2        PRC 0.2500000 1.0000000
#> 17    excel     2        ROC 0.2412000 0.9648000
#> 18    excel     2        PRC 0.2500000 1.0000000
#> 19     perf     2        ROC 0.2500000 1.0000000
#> 20     perf     2        PRC 0.2500000 1.0000000
#> 21   random     3        ROC 0.0271000 0.1084000
#> 22   random     3        PRC 0.1234462 0.4937846
#> 23  poor_er     3        ROC 0.0996000 0.3984000
#> 24  poor_er     3        PRC 0.1935903 0.7743612
#> 25  good_er     3        ROC 0.1355000 0.5420000
#> 26  good_er     3        PRC 0.2500000 1.0000000
#> 27    excel     3        ROC 0.2399000 0.9596000
#> 28    excel     3        PRC 0.2500000 1.0000000
#> 29     perf     3        ROC 0.2500000 1.0000000
#> 30     perf     3        PRC 0.2500000 1.0000000
#> 31   random     4        ROC 0.0336000 0.1344000
#> 32   random     4        PRC 0.1300030 0.5200120
#> 33  poor_er     4        ROC 0.0938000 0.3752000
#> 34  poor_er     4        PRC 0.1697297 0.6789188
#> 35  good_er     4        ROC 0.1292000 0.5168000
#> 36  good_er     4        PRC 0.2404448 0.9617793
#> 37    excel     4        ROC 0.2381000 0.9524000
#> 38    excel     4        PRC 0.2500000 1.0000000
#> 39     perf     4        ROC 0.2500000 1.0000000
#> 40     perf     4        PRC 0.2500000 1.0000000