Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves
An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object 	# of models 	# of test datasets

sscurves 	 single 	 single

mscurves 	 multiple 	 single

smcurves 	 single 	 multiple

mmcurves 	 multiple 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. part for calculation of pAUCs. auc for retieving a dataset of AUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1734000 0.4624000
#> 2    random     1        PRC 0.2039472 0.5438591
#> 3   poor_er     1        ROC 0.2914000 0.7770667
#> 4   poor_er     1        PRC 0.3695478 0.9854607
#> 5   good_er     1        ROC 0.3433000 0.9154667
#> 6   good_er     1        PRC 0.3750000 1.0000000
#> 7     excel     1        ROC 0.3750000 1.0000000
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.3959000 0.7918000
#> 2  good_er     1        PRC 0.4163517 0.8327034
#> 3  good_er     2        ROC 0.4335000 0.8670000
#> 4  good_er     2        PRC 0.4768767 0.9537535
#> 5  good_er     3        ROC 0.4193000 0.8386000
#> 6  good_er     3        PRC 0.4580020 0.9160041
#> 7  good_er     4        ROC 0.4368000 0.8736000
#> 8  good_er     4        PRC 0.4395131 0.8790263

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)

#>    modnames dsids curvetypes      paucs    spaucs
#> 1    random     1        ROC 0.01590000 0.0636000
#> 2    random     1        PRC 0.08699835 0.3479934
#> 3   poor_er     1        ROC 0.10190000 0.4076000
#> 4   poor_er     1        PRC 0.16921277 0.6768511
#> 5   good_er     1        ROC 0.15830000 0.6332000
#> 6   good_er     1        PRC 0.25000000 1.0000000
#> 7     excel     1        ROC 0.22710000 0.9084000
#> 8     excel     1        PRC 0.25000000 1.0000000
#> 9      perf     1        ROC 0.25000000 1.0000000
#> 10     perf     1        PRC 0.25000000 1.0000000
#> 11   random     2        ROC 0.03700000 0.1480000
#> 12   random     2        PRC 0.13284903 0.5313961
#> 13  poor_er     2        ROC 0.12460000 0.4984000
#> 14  poor_er     2        PRC 0.21529484 0.8611794
#> 15  good_er     2        ROC 0.12460000 0.4984000
#> 16  good_er     2        PRC 0.25000000 1.0000000
#> 17    excel     2        ROC 0.23350000 0.9340000
#> 18    excel     2        PRC 0.25000000 1.0000000
#> 19     perf     2        ROC 0.25000000 1.0000000
#> 20     perf     2        PRC 0.25000000 1.0000000
#> 21   random     3        ROC 0.03150000 0.1260000
#> 22   random     3        PRC 0.11940835 0.4776334
#> 23  poor_er     3        ROC 0.09660000 0.3864000
#> 24  poor_er     3        PRC 0.18826787 0.7530715
#> 25  good_er     3        ROC 0.12890000 0.5156000
#> 26  good_er     3        PRC 0.25000000 1.0000000
#> 27    excel     3        ROC 0.23780000 0.9512000
#> 28    excel     3        PRC 0.25000000 1.0000000
#> 29     perf     3        ROC 0.25000000 1.0000000
#> 30     perf     3        PRC 0.25000000 1.0000000
#> 31   random     4        ROC 0.01830000 0.0732000
#> 32   random     4        PRC 0.10083998 0.4033599
#> 33  poor_er     4        ROC 0.08970000 0.3588000
#> 34  poor_er     4        PRC 0.19471160 0.7788464
#> 35  good_er     4        ROC 0.14390000 0.5756000
#> 36  good_er     4        PRC 0.24360859 0.9744343
#> 37    excel     4        ROC 0.23070000 0.9228000
#> 38    excel     4        PRC 0.25000000 1.0000000
#> 39     perf     4        ROC 0.25000000 1.0000000
#> 40     perf     4        PRC 0.25000000 1.0000000