Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves

An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object	# of models	# of test datasets

sscurves	single	single

mscurves	multiple	single

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]]
)

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)
#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1600000 0.4266667
#> 2    random     1        PRC 0.2465101 0.6573603
#> 3   poor_er     1        ROC 0.2758000 0.7354667
#> 4   poor_er     1        PRC 0.3436695 0.9164521
#> 5   good_er     1        ROC 0.3255000 0.8680000
#> 6   good_er     1        PRC 0.3749362 0.9998299
#> 7     excel     1        ROC 0.3739000 0.9970667
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.4280000 0.8560000
#> 2  good_er     1        PRC 0.4640720 0.9281439
#> 3  good_er     2        ROC 0.4180000 0.8360000
#> 4  good_er     2        PRC 0.4554064 0.9108128
#> 5  good_er     3        ROC 0.3896000 0.7792000
#> 6  good_er     3        PRC 0.4313344 0.8626688
#> 7  good_er     4        ROC 0.4293000 0.8586000
#> 8  good_er     4        PRC 0.4516451 0.9032902

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
  modnames = samps[["modnames"]],
  dsids = samps[["dsids"]]
)

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)
#>    modnames dsids curvetypes      paucs    spaucs
#> 1    random     1        ROC 0.03670000 0.1468000
#> 2    random     1        PRC 0.13739692 0.5495877
#> 3   poor_er     1        ROC 0.10250000 0.4100000
#> 4   poor_er     1        PRC 0.21530678 0.8612271
#> 5   good_er     1        ROC 0.13990000 0.5596000
#> 6   good_er     1        PRC 0.25000000 1.0000000
#> 7     excel     1        ROC 0.22080000 0.8832000
#> 8     excel     1        PRC 0.24380958 0.9752383
#> 9      perf     1        ROC 0.25000000 1.0000000
#> 10     perf     1        PRC 0.25000000 1.0000000
#> 11   random     2        ROC 0.03370000 0.1348000
#> 12   random     2        PRC 0.15026903 0.6010761
#> 13  poor_er     2        ROC 0.11110000 0.4444000
#> 14  poor_er     2        PRC 0.18769264 0.7507705
#> 15  good_er     2        ROC 0.13010000 0.5204000
#> 16  good_er     2        PRC 0.25000000 1.0000000
#> 17    excel     2        ROC 0.23760000 0.9504000
#> 18    excel     2        PRC 0.25000000 1.0000000
#> 19     perf     2        ROC 0.25000000 1.0000000
#> 20     perf     2        PRC 0.25000000 1.0000000
#> 21   random     3        ROC 0.01890000 0.0756000
#> 22   random     3        PRC 0.09548669 0.3819468
#> 23  poor_er     3        ROC 0.06160000 0.2464000
#> 24  poor_er     3        PRC 0.17131663 0.6852665
#> 25  good_er     3        ROC 0.13500000 0.5400000
#> 26  good_er     3        PRC 0.25000000 1.0000000
#> 27    excel     3        ROC 0.22180000 0.8872000
#> 28    excel     3        PRC 0.25000000 1.0000000
#> 29     perf     3        ROC 0.25000000 1.0000000
#> 30     perf     3        PRC 0.25000000 1.0000000
#> 31   random     4        ROC 0.02240000 0.0896000
#> 32   random     4        PRC 0.10057776 0.4023110
#> 33  poor_er     4        ROC 0.14620000 0.5848000
#> 34  poor_er     4        PRC 0.20743814 0.8297525
#> 35  good_er     4        ROC 0.14330000 0.5732000
#> 36  good_er     4        PRC 0.24960779 0.9984312
#> 37    excel     4        ROC 0.23260000 0.9304000
#> 38    excel     4        PRC 0.25000000 1.0000000
#> 39     perf     4        ROC 0.25000000 1.0000000
#> 40     perf     4        PRC 0.25000000 1.0000000