Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves
An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object 	# of models 	# of test datasets

sscurves 	 single 	 single

mscurves 	 multiple 	 single

smcurves 	 single 	 multiple

mmcurves 	 multiple 	 multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

evalmod for generating S3 objects with performance evaluation measures. part for calculation of pAUCs. auc for retieving a dataset of AUCs.

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1637000 0.4365333
#> 2    random     1        PRC 0.1990245 0.5307319
#> 3   poor_er     1        ROC 0.2980000 0.7946667
#> 4   poor_er     1        PRC 0.3672572 0.9793525
#> 5   good_er     1        ROC 0.3491000 0.9309333
#> 6   good_er     1        PRC 0.3750000 1.0000000
#> 7     excel     1        ROC 0.3715000 0.9906667
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)

#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.3979000 0.7958000
#> 2  good_er     1        PRC 0.4170870 0.8341740
#> 3  good_er     2        ROC 0.4085000 0.8170000
#> 4  good_er     2        PRC 0.4292149 0.8584297
#> 5  good_er     3        ROC 0.4390000 0.8780000
#> 6  good_er     3        PRC 0.4673642 0.9347283
#> 7  good_er     4        ROC 0.4199000 0.8398000
#> 8  good_er     4        PRC 0.4322792 0.8645584

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)

#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.0296000 0.1184000
#> 2    random     1        PRC 0.1332009 0.5328036
#> 3   poor_er     1        ROC 0.1145000 0.4580000
#> 4   poor_er     1        PRC 0.2163178 0.8652713
#> 5   good_er     1        ROC 0.1292000 0.5168000
#> 6   good_er     1        PRC 0.2500000 1.0000000
#> 7     excel     1        ROC 0.2407000 0.9628000
#> 8     excel     1        PRC 0.2500000 1.0000000
#> 9      perf     1        ROC 0.2500000 1.0000000
#> 10     perf     1        PRC 0.2500000 1.0000000
#> 11   random     2        ROC 0.0361000 0.1444000
#> 12   random     2        PRC 0.1216187 0.4864747
#> 13  poor_er     2        ROC 0.0821000 0.3284000
#> 14  poor_er     2        PRC 0.1971065 0.7884261
#> 15  good_er     2        ROC 0.1623000 0.6492000
#> 16  good_er     2        PRC 0.2500000 1.0000000
#> 17    excel     2        ROC 0.2219000 0.8876000
#> 18    excel     2        PRC 0.2500000 1.0000000
#> 19     perf     2        ROC 0.2500000 1.0000000
#> 20     perf     2        PRC 0.2500000 1.0000000
#> 21   random     3        ROC 0.0273000 0.1092000
#> 22   random     3        PRC 0.1231919 0.4927676
#> 23  poor_er     3        ROC 0.1057000 0.4228000
#> 24  poor_er     3        PRC 0.1949931 0.7799724
#> 25  good_er     3        ROC 0.1429000 0.5716000
#> 26  good_er     3        PRC 0.2500000 1.0000000
#> 27    excel     3        ROC 0.2384000 0.9536000
#> 28    excel     3        PRC 0.2500000 1.0000000
#> 29     perf     3        ROC 0.2500000 1.0000000
#> 30     perf     3        PRC 0.2500000 1.0000000
#> 31   random     4        ROC 0.0220000 0.0880000
#> 32   random     4        PRC 0.1001821 0.4007283
#> 33  poor_er     4        ROC 0.1049000 0.4196000
#> 34  poor_er     4        PRC 0.2102833 0.8411331
#> 35  good_er     4        ROC 0.1376000 0.5504000
#> 36  good_er     4        PRC 0.2451449 0.9805796
#> 37    excel     4        ROC 0.2415000 0.9660000
#> 38    excel     4        PRC 0.2500000 1.0000000
#> 39     perf     4        ROC 0.2500000 1.0000000
#> 40     perf     4        PRC 0.2500000 1.0000000