Retrieve a data frame of pAUC scores

The auc function takes an S3 object generated by part and evalmod and retrieves a data frame with the partial AUC scores of ROC and Precision-Recall curves.

pauc(curves)

# S3 method for aucs
pauc(curves)

Arguments

curves

An S3 object generated by part and evalmod. The pauc function accepts the following S3 objects.

S3 object	# of models	# of test datasets

sscurves	single	single

mscurves	multiple	single

smcurves	single	multiple

mmcurves	multiple	multiple

See the Value section of evalmod for more details.

Value

The auc function returns a data frame with pAUC scores.

See also

Examples

##################################################
### Single model & single test dataset
###

## Load a dataset with 10 positives and 10 negatives
data(P10N10)

## Generate an sscurve object that contains ROC and Precision-Recall curves
sscurves <- evalmod(scores = P10N10$scores, labels = P10N10$labels)

## Calculate partial AUCs
sscurves.part <- part(sscurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(sscurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1       m1     1        ROC 0.3771875 0.7543750
#> 2       m1     1        PRC 0.3616708 0.7233417

##################################################
### Multiple models & single test dataset
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(1, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mscurves <- evalmod(mdat)

## Calculate partial AUCs
mscurves.part <- part(mscurves, xlim = c(0, 0.75), ylim = c(0.25, 0.75))

## Shows pAUCs
pauc(mscurves.part)
#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.1031000 0.2749333
#> 2    random     1        PRC 0.1889114 0.5037638
#> 3   poor_er     1        ROC 0.2834000 0.7557333
#> 4   poor_er     1        PRC 0.3506880 0.9351680
#> 5   good_er     1        ROC 0.3317000 0.8845333
#> 6   good_er     1        PRC 0.3733853 0.9956942
#> 7     excel     1        ROC 0.3695000 0.9853333
#> 8     excel     1        PRC 0.3750000 1.0000000
#> 9      perf     1        ROC 0.3750000 1.0000000
#> 10     perf     1        PRC 0.3750000 1.0000000

##################################################
### Single model & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "good_er")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an smcurve object that contains ROC and Precision-Recall curves
smcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
smcurves.part <- part(smcurves, xlim = c(0.25, 0.75))

## Shows pAUCs
pauc(smcurves.part)
#>   modnames dsids curvetypes     paucs    spaucs
#> 1  good_er     1        ROC 0.4197000 0.8394000
#> 2  good_er     1        PRC 0.4401356 0.8802712
#> 3  good_er     2        ROC 0.4214000 0.8428000
#> 4  good_er     2        PRC 0.4368386 0.8736771
#> 5  good_er     3        ROC 0.4254000 0.8508000
#> 6  good_er     3        PRC 0.4555504 0.9111008
#> 7  good_er     4        ROC 0.3894000 0.7788000
#> 8  good_er     4        PRC 0.4300211 0.8600422

##################################################
### Multiple models & multiple test datasets
###

## Create sample datasets with 100 positives and 100 negatives
samps <- create_sim_samples(4, 100, 100, "all")
mdat <- mmdata(samps[["scores"]], samps[["labels"]],
               modnames = samps[["modnames"]],
               dsids = samps[["dsids"]])

## Generate an mscurve object that contains ROC and Precision-Recall curves
mmcurves <- evalmod(mdat, raw_curves = TRUE)

## Calculate partial AUCs
mmcurves.part <- part(mmcurves, xlim = c(0, 0.25))

## Shows pAUCs
pauc(mmcurves.part)
#>    modnames dsids curvetypes     paucs    spaucs
#> 1    random     1        ROC 0.0249000 0.0996000
#> 2    random     1        PRC 0.1191037 0.4764146
#> 3   poor_er     1        ROC 0.0794000 0.3176000
#> 4   poor_er     1        PRC 0.1975066 0.7900263
#> 5   good_er     1        ROC 0.1452000 0.5808000
#> 6   good_er     1        PRC 0.2500000 1.0000000
#> 7     excel     1        ROC 0.2438000 0.9752000
#> 8     excel     1        PRC 0.2500000 1.0000000
#> 9      perf     1        ROC 0.2500000 1.0000000
#> 10     perf     1        PRC 0.2500000 1.0000000
#> 11   random     2        ROC 0.0365000 0.1460000
#> 12   random     2        PRC 0.1365326 0.5461305
#> 13  poor_er     2        ROC 0.1344000 0.5376000
#> 14  poor_er     2        PRC 0.2350818 0.9403271
#> 15  good_er     2        ROC 0.1463000 0.5852000
#> 16  good_er     2        PRC 0.2500000 1.0000000
#> 17    excel     2        ROC 0.2317000 0.9268000
#> 18    excel     2        PRC 0.2500000 1.0000000
#> 19     perf     2        ROC 0.2500000 1.0000000
#> 20     perf     2        PRC 0.2500000 1.0000000
#> 21   random     3        ROC 0.0430000 0.1720000
#> 22   random     3        PRC 0.1707759 0.6831036
#> 23  poor_er     3        ROC 0.1062000 0.4248000
#> 24  poor_er     3        PRC 0.2230720 0.8922881
#> 25  good_er     3        ROC 0.1538000 0.6152000
#> 26  good_er     3        PRC 0.2500000 1.0000000
#> 27    excel     3        ROC 0.2392000 0.9568000
#> 28    excel     3        PRC 0.2500000 1.0000000
#> 29     perf     3        ROC 0.2500000 1.0000000
#> 30     perf     3        PRC 0.2500000 1.0000000
#> 31   random     4        ROC 0.0335000 0.1340000
#> 32   random     4        PRC 0.1281691 0.5126763
#> 33  poor_er     4        ROC 0.0743000 0.2972000
#> 34  poor_er     4        PRC 0.1648234 0.6592935
#> 35  good_er     4        ROC 0.1490000 0.5960000
#> 36  good_er     4        PRC 0.2500000 1.0000000
#> 37    excel     4        ROC 0.2286000 0.9144000
#> 38    excel     4        PRC 0.2500000 1.0000000
#> 39     perf     4        ROC 0.2500000 1.0000000
#> 40     perf     4        PRC 0.2500000 1.0000000